% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-degree-groups.R
\name{assign_degree_groups}
\alias{assign_degree_groups}
\title{Assign groups to a vector of degree angles}
\usage{
assign_degree_groups(degrees, min_degrees = 10, ...)
}
\arguments{
\item{degrees}{\code{numeric} vector of degree angles, in any order.}

\item{min_degrees}{\code{numeric} value indicating the threshold for
grouping values in \code{degrees} together.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} with rows in the same order as input \code{degrees},
with column 'group' indicating the group assignment as an \code{integer}.
}
\description{
Assign groups to a vector of degree angles
}
\examples{
assign_degree_groups(c(5, 10, 355, 350))

assign_degree_groups(c(5, 10, 355, 180, 350, 340, 174))
}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{check_systemfonts_family}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_degrees_clockwise}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{render_venndir_footnotes}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{subset_systemfonts}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
