% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{bbox_JamPolygon}
\alias{bbox_JamPolygon}
\title{Bounding box for JamPolygon}
\usage{
bbox_JamPolygon(jp, ...)
}
\arguments{
\item{jp}{\code{JamPolygon} object}
}
\description{
Bounding box for JamPolygon
}
