% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{bbox_polygon_list}
\alias{bbox_polygon_list}
\title{Bounding box for polygon list}
\usage{
bbox_polygon_list(polygon_list, ...)
}
\description{
Bounding box for polygon list
}
\examples{
bbox_polygon_list(polygon_list)

}
