% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{buffer_JamPolygon}
\alias{buffer_JamPolygon}
\title{Apply buffer outside or inside JamPolygon}
\usage{
buffer_JamPolygon(
  jp,
  buffer = -0.5,
  steps = 50,
  relative = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{jp}{\code{JamPolygon} with one or more polygons. When multiple polygons
are provided, they are combined with \code{union_JamPolygon()} so that
one overall buffer can be provided.}

\item{buffer}{\code{numeric} buffer, where negative values cause the polygon
to be reduced in size.}
}
\description{
Apply buffer outside or inside JamPolygon
}
