% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{check_JamPolygon}
\alias{check_JamPolygon}
\title{Check JamPolygon object}
\usage{
check_JamPolygon(object)
}
\arguments{
\item{object}{\code{JamPolygon} object}
}
\description{
Check whether a JamPolygon object is valid.
}
\details{
It requires:
\itemize{
\item \code{slotName(object)} to contain \code{"polygons"}
\item \code{object@polygons} to inherit \code{"data.frame"}
\item \code{colnames(object@polygons)} to contain \code{c("name", "x", "y")}
}

General guidance for JamPolygon objects:
\itemize{
\item Empty polygons can be represented as one row in the \code{data.frame}
which contains \code{NULL} or \code{numeric(0)} x,y coordinates.
\itemize{
\item Empty polygons can be the result of \code{intersect_JamPolygon()}
that finds no suitable intersect.
\item Note that downstream processing, and plotting, must account for
empty polygons and handle or ignore them accordingly.
}
}
}
