% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-systemfonts.R
\name{check_systemfonts_family}
\alias{check_systemfonts_family}
\title{Check systemfonts family for valid freetype access}
\usage{
check_systemfonts_family(family = "Arial", ...)
}
\arguments{
\item{family}{\code{character} vector with font family names}

\item{...}{additional arguments are ignored.}
}
\value{
\code{integer} vector with the number of valid font info rows
returned by \code{systemfonts::font_info()}, for each entry provided
in \code{family}.
}
\description{
Check systemfonts family for valid freetype access
}
\examples{
check_systemfonts_family(c("Noto Sans Syriac", "Skia", "PingFang TC", "Arial"))

check_systemfonts_family(c("Arial", "Arial"))

}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{hexsticker_venndir}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{subset_systemfonts}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
