% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndir-class.R
\name{check_Venndir}
\alias{check_Venndir}
\title{Check Venndir object}
\usage{
check_Venndir(object)
}
\arguments{
\item{object}{\code{Venndir} object}
}
\description{
Check Venndir object integrity
}
\details{
\subsection{Todo:}{
\itemize{
\item Write additional validation checks.
}
}
}
\seealso{
Other venndir s4: 
\code{\link{Venndir-class}},
\code{\link{plot,Venndir,ANY-method}}
}
\concept{venndir s4}
