% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-collapse-im.R
\name{collapse_im}
\alias{collapse_im}
\title{Collapse incidence matrix using row groups}
\usage{
collapse_im(
  im,
  row_groups = NULL,
  logic = c("majority-hit", "majority"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{im}{\code{numeric} matrix with columns for each set}

\item{row_groups}{\code{character} or \code{factor} with row groupings.}

\item{logic}{\code{character} logic to use, default 'majority-hit'.
\itemize{
\item \code{"majority-hit"}: uses the majority winner among non-zero values.
\item \code{"majority"}: uses the majority winner including non-zero and zero.
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} matrix
}
\description{
Collapse incidence matrix using row groups, for example when
converting probe-level, transcript-level, peptide-level data
to gene-level data.
}
\details{
This function is a simple converted for incidence matrix data,
taking the "majority-hit" for each row grouping. The most common
scenario is to group rows by gene, in order to summarize the
observed changes at gene level, when the original data may
contain multiple possible measurements for each gene.

The default logic assumes that \emph{any} observed statistical hit for
a gene is sufficient evidence to implicate that gene as a "hit",
even if other potential measurements for the same gene did not
meet the statistical criteria used, as relevant to the platform
technology.
}
\examples{
im <- cbind(A=c(-1, -1, 0, 1, 1, 1, -1, 0, 0, 1, 1, 0),
   B=c(-1, -1, -1, 1, 1, 0, -1, 0, 0, 1, 1, 1),
   C=c(-1, -1, -1, 1, 1, 0, -1, 0, 0, 0, 0, 0));
row_groups <- rep(c("a", "b", "c"), c(6, 3, 3))

# default logic returns the majority non-zero value when present
new_im <- collapse_im(im, row_groups)
new_im

# majority logic will prioritize "0" when it is the majority
# (not recommended for most gene-based data)
new_im2 <- collapse_im(im, row_groups, logic="majority")
new_im2

# more detail
imdf <- data.frame(im, row_groups,
   new_im[match(row_groups, rownames(new_im)), ])
split(imdf, imdf$row_groups)
}
\seealso{
Other venndir conversion: 
\code{\link{counts2setlist}()},
\code{\link{im2list}()},
\code{\link{im_value2list}()},
\code{\link{list2im_opt}()},
\code{\link{list2im_value}()},
\code{\link{overlaplist2setlist}()},
\code{\link{signed_counts2setlist}()}
}
\concept{venndir conversion}
