% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-conversion.R
\name{counts2setlist}
\alias{counts2setlist}
\title{venndir conversion from overlap counts to setlist}
\usage{
counts2setlist(x, sep = "&", ...)
}
\value{
\code{list} where the list names are the names of each set,
and values of each list element is a vector of items.
The items are artificial labels used for convenience.
}
\description{
venndir conversion from overlap counts to setlist
}
\details{
This function takes input in the form of named \code{vector}
with overlap counts, which is the input to similar Venn
functions in \code{eulerr::euler()}, and is called "expressionInput"
when used with \code{upsetr::fromExpression()}.
}
\examples{
# example from eulerr::euler()
combo <- c(A = 2,
   B = 2,
   C = 2,
   "A&B" = 2,
   "A&C" = 1,
   "B&C" = 1,
   "A&B&C" = 4)
setlist <- counts2setlist(combo)
setlist;

# Venn diagram
venndir(setlist)

# Proportional Venn (Euler) diagram
venndir(setlist, proportional=TRUE)

}
\seealso{
Other venndir conversion: 
\code{\link{im2list}()},
\code{\link{im_value2list}()},
\code{\link{list2im_opt}()},
\code{\link{list2im_value}()},
\code{\link{overlaplist2setlist}()}
}
\concept{venndir conversion}
