% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-curate.R
\name{curate_venn_labels}
\alias{curate_venn_labels}
\title{Curate Venn labels}
\usage{
curate_venn_labels(
  x,
  type = c("sign", "color"),
  curate_df = NULL,
  unicode = TRUE,
  ...
)
}
\description{
Curate Venn labels
}
\seealso{
Other venndir utility: 
\code{\link{eulerr2polys}()},
\code{\link{get_venn_shapes}()},
\code{\link{make_venn_test}()},
\code{\link{nudge_sp}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()}
}
\concept{venndir utility}
