% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-angles.R
\name{diff_degrees}
\alias{diff_degrees}
\title{Angular difference in degrees}
\usage{
diff_degrees(degree1, degree2 = NULL, ...)
}
\arguments{
\item{degree1}{\code{numeric} vector that contains angles in degrees.}

\item{degree2}{\code{numeric} vector that contains angles in degrees,
or \code{NULL} when the input should only consider \code{degree1}.}

\item{...}{additional arguments are ignored.}
}
\description{
Angular difference in degrees
}
\details{
This function simply returns the difference between two
angles in degrees. When \code{degree1} and \code{degree2} are vectors,
this function operates on the full set in one step.

When \code{degree1} is a vector, and \code{degree2} is \code{NULL}, this
function calculates the difference of each pair in \code{degree1},
in the order supplied.
}
\examples{
diff_degrees(5, 355)
diff_degrees(355, 5)
diff_degrees(-10, 5)
diff_degrees(-10, 355)
diff_degrees(180, 361)

}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{display_angles}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{intersect_polygons}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_sp}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()}
}
\concept{venndir spatial}
