% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-angles.R
\name{display_angles}
\alias{display_angles}
\title{Display degree angles around a unit circle}
\usage{
display_angles(x, add = FALSE, col = "darkorange", lwd = 2, ...)
}
\description{
Display degree angles around a unit circle
}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{intersect_polygons}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_sp}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()}
}
\concept{venndir spatial}
