% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{eulerr2polys}
\alias{eulerr2polys}
\title{Convert eulerr() output to polygons}
\usage{
eulerr2polys(x)
}
\description{
Convert eulerr() output to polygons
}
\examples{
if (require(eulerr)) {
}

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{get_venn_shapes}()},
\code{\link{make_venn_test}()},
\code{\link{nudge_sp}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()}
}
\concept{venndir utility}
\section{@import sp}{
}
