% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{eulerr_to_JamPolygon}
\alias{eulerr_to_JamPolygon}
\title{Convert eulerr output to JamPolygon}
\usage{
eulerr_to_JamPolygon(x)
}
\value{
\code{JamPolygon} object
}
\description{
Convert eulerr output to JamPolygon
}
