% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-label-outside-jp.R
\name{farthest_point_JamPolygon}
\alias{farthest_point_JamPolygon}
\title{Get the farthest polygon point from a reference point}
\usage{
farthest_point_JamPolygon(x, jp, ...)
}
\arguments{
\item{x}{\code{numeric} matrix with columns \code{"x"} and \code{"y"}}

\item{jp}{\code{JamPolygon}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} matrix with columns \code{"x"} and \code{"y"}
}
\description{
Get the farthest polygon point from a reference point
}
\seealso{
Other JamPolygon: 
\code{\link{area_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
