% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{get_largest_polygon}
\alias{get_largest_polygon}
\title{Largest polygon in a SpatialPolygons object}
\usage{
get_largest_polygon(sp, ...)
}
\arguments{
\item{sp}{object with class \code{sp::SpatialPolygons} that may
contain one or more closed polygons.}

\item{...}{additional arguments are ignored.}
}
\description{
Largest polygon in a SpatialPolygons object
}
\details{
This function returns the largest polygon in a
\code{sp::SpatialPolygons} object, notably when there
are multiple polygons contained in one object.
The function calculates the largest area using
\code{rgeos::gArea()}.

If two polygons have identical area, the first
polygon is returned.

This function calls a helper function \code{sp::disaggregate()}
which creates one \code{sp::SpatialPolygons} object
for each individual polygon in the source object.
}
\examples{
circles <- get_venn_shapes(c(A=1, B=2, "A&B"=3), proportional=TRUE)
col3 <- c("#FF000077", "#0000FF77")
sp::plot(circles)

circles_u12 <- union_polygons(circles[1:2]);
circles_i12 <- intersect_polygons(circles[1:2]);

# one SpatialPolygons object with multiple polygon pieces
circles_d12 <- rgeos::gDifference(circles_u12, circles_i12)
plot(circles_d12, add=TRUE, col="gold", lwd=3)

circles_d12_largest <- get_largest_polygon(circles_d12);
plot(circles_d12_largest, add=TRUE, border="red", lwd=5);

}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{get_sp_buffer}()},
\code{\link{intersect_polygons}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_sp}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()}
}
\concept{venndir spatial}
