% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{get_largest_polygon_list}
\alias{get_largest_polygon_list}
\title{Largest polygon in a polygon list}
\usage{
get_largest_polygon_list(polygon_list, ...)
}
\arguments{
\item{polygon_list}{\code{list} with \code{"x"} and \code{"y"} elements with
polygon coordinates.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} with polygon coordinates \code{"x"} and \code{"y"}
}
\description{
Largest polygon in a polygon list
}
\details{
This function returns the largest polygon in a polygon list,
intended when there are multiple polygons contained in one object.

If two polygons have identical area, the first
polygon is returned.
\subsection{Todo:}{
\itemize{
\item Verify correct output when polygon(s) have holes.
}
}
}
\examples{
counts <- c(A=1, B=2, `A&B`=3, C=4)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
largest_poly <- get_largest_polygon_list(polygon_list)
plot_polygon_list(polygon_list, col=colorjam::rainbowJam(3, alpha=0.5))
plot_polygon_list(largest_poly, add=TRUE, border="red", lwd=3)


}
\seealso{
Other venndir polygons: 
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_polygon_buffer}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_coord_list}()}
}
\concept{venndir polygons}
