% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-shapes.R
\name{get_venn_polygon_shapes}
\alias{get_venn_polygon_shapes}
\title{Get Venn shapes as polygon_list}
\usage{
get_venn_polygon_shapes(
  counts,
  proportional = FALSE,
  sep = "&",
  circles_only = FALSE,
  circle_nudge = NULL,
  rotate_degrees = 0,
  return_type = c("polygon_list", "JamPolygon"),
  ...
)
}
\arguments{
\item{counts}{\code{integer} vector whose names represent set overlaps,
where set names are separated by \code{sep} delimiter.}

\item{proportional}{\code{logical} indicating whether to create proportional
circles, where \code{proportional=FALSE} creates standard Venn diagram,
and \code{proportional=TRUE} creates a Euler diagram.}

\item{sep}{\code{character} delimiter used to separate set names in
\code{names(counts)}.}

\item{circles_only}{\code{logical} indicating whether to force Venn
4-way diagram to use only circles; or passed to \code{eulerr::euler()}
to force it to return circles instead of allowing ellipse shapes.}

\item{circle_nudge}{\code{list} of \code{numeric} vectors each length 2, whose
names match set names derived from \code{counts}. For example if
\code{counts=c(set_A=5, set_B=10, "setA&set_B"=3)}, then to nudge
the \code{set_A} circle, use \code{circle_nudge=list(set_A=c(1, 0))}.
This argument is intended to allow manipulation of specific
circle or ellipse positions for aesthetic effects. Particularly
for proportional Euler diagrams, sometimes the algorithm places
circles in non-ideal locations}

\item{rotate_degrees}{\code{numeric} value indicating rotation in degrees
for the entire set of shapes. This argument is intended to
change the overall orientation, for example so that certain
sets are at the top.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} in polygon_list format with \code{"x"} and \code{"y"} elements,
or when \code{return_type="JamPolygon"} it returns \code{JamPolygon}.
}
\description{
Get Venn shapes as polygon_list
}
\details{
This function takes a Venn overlap counts and creates
corresponding circles or ellipses that represent
either a Venn diagram, or proportional Venn (Euler)
diagram.

For non-proportional Venn diagrams, this function accepts
up to 5 sets, although the 5-way Venn diagram is not
visually intuitive.

For proportional Euler diagrams, this function simply passes
the count vector to \code{eulerr::euler()} and returns the output.
That function accepts more sets, however not all overlaps may
be represented in the output.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3, C=4)
venn_colors <- colorjam::rainbowJam(3, alpha=0.5);

venn_polygon_list <- get_venn_polygon_shapes(counts)
plot_polygon_list(venn_polygon_list, col=venn_colors)

venn_polygon_list <- get_venn_polygon_shapes(counts, proportional=TRUE)
plot_polygon_list(venn_polygon_list, col=venn_colors)

# TODO: examples showing circle_nudge, rotate_degrees
jpdf <- get_venn_polygon_shapes(counts, return_type="JamPolygon")

}
\seealso{
Other venndir polygons: 
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_polygon_buffer}()},
\code{\link{intersect_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_coord_list}()}
}
\concept{venndir polygons}
