% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-curate-df.R
\name{get_venndir_curate_df}
\alias{get_venndir_curate_df}
\title{Define venndir curate data.frame}
\usage{
get_venndir_curate_df(unicode = TRUE, use_color = TRUE, curate_df = NULL, ...)
}
\arguments{
\item{unicode}{\code{logical} default TRUE, whether to use unicode output
values. Optionally \code{unicode=2} will return an alternative set of
Unicode values.}

\item{use_color}{\code{logical} default TRUE, whether to apply color,
a convenient method to use dark grey.}

\item{curate_df}{\code{data.frame} with optional user-defined values,
which are used in place of corresponding default values in the
'from' column.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} with colnames:
\itemize{
\item 'from' - \code{character} sign to match
\item 'color' - \code{character} color associated with matched values
\item 'sign' - \code{character} replacement of the matched value in 'from'
\item 'hide_singlet' - \code{logical} indicating whether to hide the signed
count value when the overlap is a singlet, for example "agreement"
has no meaning when there is only one set involved.
}
}
\description{
Define venndir curate data.frame
}
\details{
This function defines the \code{data.frame} used for the argument
\code{curate_df} in \code{venndir()}. The purpose is to make it more
convenient to review and potentially modify the default values.
}
\examples{
# default
get_venndir_curate_df()

# alternate Unicode
get_venndir_curate_df(unicode=2)

# non-unicode
get_venndir_curate_df(unicode=FALSE)

# custom curate_df
my_curate_df <- data.frame(from="1", sign="^", color="red")
get_venndir_curate_df(unicode=TRUE, curate_df=my_curate_df)
}
\seealso{
Other venndir advanced: 
\code{\link{curate_venn_labels}()},
\code{\link{modify_venndir_overlap}()},
\code{\link{nudge_venndir_label}()}
}
\concept{venndir advanced}
