% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-gridtext.R
\name{gridtext_make_outer_box}
\alias{gridtext_make_outer_box}
\title{Custom version of gridtext make_outer_box}
\usage{
gridtext_make_outer_box(
  vbox_inner,
  width,
  height,
  x,
  y,
  halign,
  valign,
  hjust,
  vjust,
  rot,
  margin_pt,
  padding_pt,
  r_pt,
  box_gp
)
}
\description{
Custom version of gridtext make_outer_box
}
\details{
This function is intended to provide support for vectorized \code{padding},
so that each label can have its own distinct padding on all four sides.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{gridtext_richtext_grob}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_venndir_label}()},
\code{\link{print_color_df}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{shrink_df}()},
\code{\link{sp_percent_area}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_legender}()}
}
\concept{venndir utility}
