% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-gridtext.R
\name{gridtext_richtext_grob}
\alias{gridtext_richtext_grob}
\title{Custom gridtext richtext grob}
\usage{
gridtext_richtext_grob(
  text,
  x = grid::unit(0.5, "npc"),
  y = grid::unit(0.5, "npc"),
  hjust = 0.5,
  vjust = 0.5,
  halign = hjust,
  valign = vjust,
  rot = 0,
  default.units = "npc",
  margin = grid::unit(c(0, 0, 0, 0), "pt"),
  padding = grid::unit(c(0, 0, 0, 0), "pt"),
  r = grid::unit(0, "pt"),
  align_widths = FALSE,
  align_heights = FALSE,
  name = NULL,
  gp = grid::gpar(),
  box_gp = grid::gpar(col = NA),
  vp = NULL,
  use_markdown = TRUE,
  debug = FALSE
)
}
\description{
Custom gridtext richtext grob
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{gridtext_make_outer_box}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_venndir_label}()},
\code{\link{print_color_df}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{shrink_df}()},
\code{\link{sp_percent_area}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_legender}()}
}
\concept{venndir utility}
