% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-grob-assembly.R
\name{grobs_stack}
\alias{grobs_stack}
\title{Stack grobs top to bottom}
\usage{
grobs_stack(
  grobs,
  name = NULL,
  ref_grob = 1,
  use_y = NULL,
  verbose = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{grobs}{one of the following forms of \code{grid} graphical objects:
\itemize{
\item \code{list} containing grid grobs, or \code{gTree} objects.
\item single \code{gTree} objects, which is broken into a \code{list} of components
\item \code{gList} containing grid grobs or \code{gTree} objects.
}}

\item{ref_grob}{\code{integer} intended to define the reference object,
and refers to the object in order they are defined in the \code{gTree}.}

\item{use_y}{\code{unit} object, experimental, used instead of \code{ref_grob}
to define the starting position.}

\item{verbose}{\code{logical} indicating whether to print verbose output}

\item{debug}{\code{logical} with experimental debug output}

\item{...}{additional arguments are ignored.}
}
\description{
Stack grobs top to bottom
}
\details{
This approach uses proper xext,yext min and max values from \code{grobs_exts()}
}
\seealso{
Other grobs: 
\code{\link{grobs_exts}()},
\code{\link{grobs_tile}()},
\code{\link{grobs_xalign}()},
\code{\link{grobs_yalign}()}
}
\concept{grobs}
\keyword{internal}
