% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-grob-assembly.R
\name{grobs_xalign}
\alias{grobs_xalign}
\title{Left-align grobs}
\usage{
grobs_xalign(
  grobs,
  name = NULL,
  xalign = c("left", "center", "right", "tile", "none"),
  ref_grob = 1,
  use_x = NULL,
  verbose = FALSE,
  debug = FALSE,
  ...
)
}
\description{
Left-align grobs
}
\seealso{
Other grobs: 
\code{\link{grobs_exts}()},
\code{\link{grobs_stack}()},
\code{\link{grobs_tile}()},
\code{\link{grobs_yalign}()}
}
\concept{grobs}
\keyword{internal}
