% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-grob-assembly.R
\name{grobs_yalign}
\alias{grobs_yalign}
\title{Top-bottom grobs}
\usage{
grobs_yalign(
  grobs,
  name = NULL,
  yalign = c("top", "middle", "bottom"),
  ref_grob = 1,
  use_y = NULL,
  verbose = FALSE,
  debug = FALSE,
  ...
)
}
\description{
Top-bottom grobs
}
\seealso{
Other grobs: 
\code{\link{grobs_exts}()},
\code{\link{grobs_stack}()},
\code{\link{grobs_tile}()},
\code{\link{grobs_xalign}()}
}
\concept{grobs}
\keyword{internal}
