% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{has_point_in_JamPolygon}
\alias{has_point_in_JamPolygon}
\title{Determine if a point is inside any JamPolygon}
\usage{
has_point_in_JamPolygon(x, jp, apply_holes = TRUE, ...)
}
\arguments{
\item{x}{\code{list} with \code{"x"},\code{"y"} with \code{numeric} values, representing
one or more points.}

\item{jp}{\code{JamPolygon}}

\item{apply_holes}{\code{logical} (default TRUE) whether internal polygon
holes are applied, which requires a point in a solid portion of
the polygon.
When \code{apply_holes=FALSE} a point must only be contained within the
outer boundary of the polygon.}

\item{...}{additional arguments are passed to \code{point_in_JamPolygon()}}
}
\value{
\code{logical} vector with one result per point \code{x}, where
\itemize{
\item \code{TRUE} indicates the point overlaps at least one polygon,
\item \code{FALSE} indicates the point does not overlap any polygon.
\item Note that a point contained in a polygon "hole" is expected
to return \code{FALSE} when \code{apply_holes=TRUE} (default).
}
}
\description{
Determine if a point is inside any JamPolygon
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon,ANY,ANY,ANY-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
