% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-im.R
\name{im_value2list}
\alias{im_value2list}
\title{Value incidence matrix to list}
\usage{
im_value2list(x, empty = c(NA, "", 0), ...)
}
\value{
\code{list} whose names are set names derived from \code{colnames(x)},
and where each \code{vector} is named using \code{rownames(x)} and the
values from \code{x}, for non-empty values.
}
\description{
Value incidence matrix to list
}
\details{
This function is the reciprocal to \code{list2im_value()}.

A value incidence matrix is a matrix whose non-empty values
are retained in each item vector, where items are stored as
vector names.

This function is most commonly used with signed values
\code{c(-1, 0, 1)}, to indicate direction where \code{1} is \code{up},
\code{-1} is \code{down}, and \code{0} is \verb{not changed}. In this case \code{0}
which is considered empty by default.
}
\examples{
setlist <- make_venn_test(100, 3, do_signed=TRUE)
ims <- list2im_value(setlist);
print(head(ims));

# convert back to list
im_value2list(ims);

# make a simple character vector list
setlistv <- lapply(setlist, function(i){
   j <- letters[i+3];
   names(j) <- names(i);
   j;
})
imv <- list2im_value(setlistv);
print(head(imv));
im_value2list(imv);

}
\seealso{
Other venndir conversion: 
\code{\link{counts2setlist}()},
\code{\link{im2list}()},
\code{\link{list2im_opt}()},
\code{\link{list2im_value}()},
\code{\link{overlaplist2setlist}()}
}
\concept{venndir conversion}
