% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{intersect_polygon_list}
\alias{intersect_polygon_list}
\title{Intersect one or more polygons}
\usage{
intersect_polygon_list(polygon_list, new_name = NULL, ...)
}
\arguments{
\item{polygon_list}{\code{list} object that contains one or more polygons.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{list} of polygons
}
\description{
Intersect one or more polygons
}
\details{
This function takes a \code{list} of polygons and iteratively
calls \code{polyclip::polyclip(A, B, op="intersect")} to produce the intersect
across one or more polygons, which otherwise only works with two
polygons.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)

circle_intersect <- intersect_polygon_list(polygon_list);
jamba::ssdim(circle_intersect)
circle_colors <- colorjam::rainbowJam(2);
plot_polygon_list(polygon_list, col=circle_colors, main="intersect")
plot_polygon_list(circle_intersect, col="#FFDD0088", border="red", lwd=3, add=TRUE)

}
\seealso{
Other venndir polygons: 
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_polygon_buffer}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_coord_list}()}
}
\concept{venndir polygons}
