% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{intersect_polygons}
\alias{intersect_polygons}
\title{Intersect one or more polygons}
\usage{
intersect_polygons(sp, ...)
}
\description{
Intersect one or more polygons
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{make_venn_test}()},
\code{\link{nudge_sp}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()}
}
\concept{venndir utility}
