% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-label-outside-jp.R
\name{label_segment_JamPolygon}
\alias{label_segment_JamPolygon}
\title{Define a label segment for JamPolygon}
\usage{
label_segment_JamPolygon(
  x0,
  x1,
  y0,
  y1,
  jp,
  return_class = c("point", "matrix"),
  buffer = 0,
  plot_debug = FALSE,
  relative = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x0, y0}{\code{numeric} coordinate of starting point, assumed to be
outside the polygon \code{jp}.}

\item{x1, y1}{\code{numeric} coordinate of end point, assumed to be
inside the polygon \code{jp}. If not inside the polygon, the nearest
polygon point is used.}

\item{buffer}{\code{numeric} buffer applied to the polygon, where negative
would shrink the polygon, and positive would enlarge it. Default 0.
Negative values are useful to have the line segment end
just inside the polygon, by shrink the polygon, drawing a line to
the edge of the smaller polygon.}

\item{plot_debug}{\code{logical} default FALSE, whether to plot debug
information for visual review. Not implemented.}

\item{relative}{\code{logical} default TRUE, whether the \code{buffer} is fraction
of max plot dimensions.}

\item{verbose}{\code{logical} whether to print verbose output.}
}
\value{
\code{numeric} matrix with columns
\itemize{
\item x,y \code{numeric} with x,y coordinates of the end point, where
the input x0,y0 are expected to represent the starting coordinates.
\item adjx,adjy \code{numeric} values used to orient text relative to the
line, assuming the text would be positioned at the x0,y0
coordinates.
\item When \code{return_class='matrix'} the output contains two rows, the first
row is the starting point (given as x0, y0) and the second row is the
point inside the \code{jp} polygon.
}
}
\description{
Define a label segment for JamPolygon
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon,ANY,ANY,ANY-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
