% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{labelr_JamPolygon}
\alias{labelr_JamPolygon}
\title{Define label positions for JamPolygon using polylabelr::poi()}
\usage{
labelr_JamPolygon(jp, precision = 1, add_to_jp = FALSE, ...)
}
\arguments{
\item{jp}{\code{JamPolygon}}

\item{precision}{\code{numeric} passed to \code{polylabelr::poi()}}

\item{add_to_jp}{\code{logical} indicating whether to add \code{"label_x"},
\code{"label_y"} into the \code{jp@polygons} \code{data.frame}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{matrix} with columns \code{"x"}, \code{"y"}, \code{"dist"}
}
\description{
Define label positions for JamPolygon using polylabelr::poi()
}
\details{
This function is a simple wrapper for \code{polylabelr::poi()}, which
recognizes holes inside polygons.
\subsection{Todo:}{
\itemize{
\item Consider basic support for non-overlapping label positions.
\enumerate{
\item First pass might be to use non-overlapping regions of a polygon
among a set of polygons.
\item Second pass would be to use the region with fewest overlaps
among other polygons in the set, to define the label position.
This rule could serve to solve (1) as well.
}
}
}
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{[,JamPolygon-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
