% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{labelr_polygon_list}
\alias{labelr_polygon_list}
\title{Calculate polygon label positions using Pole of Inaccessibility}
\usage{
labelr_polygon_list(polygon_list, add_labels = FALSE, ...)
}
\arguments{
\item{polygon_list}{\code{list} containing elements \code{"x"} and \code{"y"} each
with \code{numeric} vectors, or \code{list} of \code{numeric} vectors.}

\item{add_labels}{\code{logical} indicating whether to plot the labels
using \code{text()}}

\item{...}{additional arguments are passed to \code{text()} when
\code{add_labels=TRUE}}
}
\description{
Calculate polygon label positions using Pole of Inaccessibility, otherwise
known as the Visual Center.
}
\details{
This function is a wrapper for \code{polylabelr::poi()} except that it
is applied to a \code{list} of polygons individually.

When any one polygon is composed of two smaller polygon components,
as encoded with a nested list of coordinates,
first the polygons are combined using \code{union_polygon_list()}.
If the result is a single polygon, that is used to define the
label position. If the result is multiple separate polygon
components, the largest polygon component is used to find the label.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3, C=5, `B&C`=2, `A&C`=2, `A&B&C`=1)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)

# default is to label each polygon in its center
plot_polygon_list(polygon_list,
   col=colorjam::rainbowJam(length(polygon_list), alpha=0.5))
labelr_polygon_list(polygon_list, add_labels=TRUE)

# create unique polygons for each label
A_only <- minus_polygon_list(polygon_list, new_name="A_only");
B_only <- minus_polygon_list(polygon_list[c(2,1,3)], new_name="B_only");
C_only <- minus_polygon_list(polygon_list[c(3,1,2)], new_name="C_only");

plot_polygon_list(polygon_list,
   col=colorjam::rainbowJam(length(polygon_list), alpha=0.5))
ABC_only <- c(A_only, B_only, C_only);
polygon_list_labelr(ABC_only, add_labels=TRUE)

# label ABC intersection
ABC_int <- intersect_polygon_list(polygon_list[c(1,2,3)], new_name="ABC");
plot_polygon_list(ABC_int, add=TRUE, col="gold")
polygon_list_labelr(ABC_int, add_labels=TRUE)

# label AB intersection
AB_only <- minus_polygon_list(
   c(intersect_polygon_list(polygon_list[c(1,2)], new_name="BC_only"),
      polygon_list[3]))
plot_polygon_list(AB_only, add=TRUE, col="darkviolet")
polygon_list_labelr(AB_only, add_labels=TRUE, col="white")

# label BC intersection
BC_only <- minus_polygon_list(
   c(intersect_polygon_list(polygon_list[c(2,3)], new_name="BC_only"),
      polygon_list[1]))
plot_polygon_list(BC_only, add=TRUE, col="skyblue")
polygon_list_labelr(BC_only, add_labels=TRUE)

}
