% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-im.R
\name{list2im_value}
\alias{list2im_value}
\title{Convert list to a value incidence matrix}
\usage{
list2im_value(
  setlist,
  empty = NULL,
  do_sparse = FALSE,
  force_sign = FALSE,
  ...
)
}
\arguments{
\item{setlist}{\code{list} of vectors}

\item{empty}{default single value used for empty/missing entries,
the default \code{empty=0} uses zero for entries not present.
Another alternative is \code{NA}.}

\item{do_sparse}{\code{logical} indicating whether to coerce the output
to sparse matrix class \code{"ngCMatrix"} from the Matrix package.
The default is \code{FALSE}.}

\item{...}{additional arguments are ignored.}

\item{coerce_sign}{\code{logical} indicating whether to coerce numeric
vector values to the sign. When \code{coerce_sign=FALSE} the vector
values are stored directly. When \code{coerce_sign=TRUE} the signs of
the vector values are stored.}
}
\value{
\code{Matrix} object that contains signed direction encoded
as \code{c(-1, 0, 1)} values.
}
\description{
Optimized conversion of list to a value incidence matrix
}
\details{
This function converts a list of named vectors into
an incidence matrix with value for each entry
(row) present in each input list (column). This output
is called a "value incidence matrix" because the
value itself is included in the matrix as opposed to
a true incidence matrix that represents only \code{TRUE}
or \code{FALSE} (\code{1} or \code{0}) at each position.

The rownames of the output matrix represent items,
encoded by the vector names. The colnames of the output
matrix represent the list names.

The default value in the output matrix is \code{0} for
a numeric matrix, and \code{""} for a character matrix,
based on the input vector classes.

To override this behavior, use the argument \code{empty}.
For example, it may be useful to encode missing entries
as \code{NA}, which means "this entry was not observed", and
store true values of \code{0} to indicate "this entry was observed
and its value was \code{0} zero". In this case use \code{empty=NA}.

This behavior can be useful with gene expression data,
when a particular gene may not be observed in all data sets.
}
\examples{
setlist <- make_venn_test(100, 3, do_signed=TRUE)
imv <- list2im_value(setlist);
print(head(imv));

# convert back to list
im_value2list(imv);

# example showing empty=NA
setlist2 <- make_venn_test(20, 3, sizes=c(12, 14, 8), do_signed=TRUE)
setlist2 <- lapply(setlist2, function(i){
   i[] <- sample(c(-1, 0, 1),
      replace=TRUE,
      size=length(i));
   i
})
imv2 <- list2im_value(setlist2, empty=NA);
imv2;

# to convert back to list, define empty=NA so 0 is not considered as empty
im_value2list(imv2, empty=NA);

# make a simple character vector list
setlistv <- lapply(setlist, function(i){
   j <- letters[i+3];
   names(j) <- names(i);
   j;
})
imv <- list2im_value(setlistv);
print(head(imv));

# convert back to list of character vectors
im_value2list(imv);

}
\seealso{
Other venndir conversion: 
\code{\link{collapse_im}()},
\code{\link{counts2setlist}()},
\code{\link{im2list}()},
\code{\link{im_value2list}()},
\code{\link{list2im_opt}()},
\code{\link{overlaplist2setlist}()},
\code{\link{signed_counts2setlist}()}
}
\concept{venndir conversion}
