% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-printcolordf.R
\name{make_color_contrast}
\alias{make_color_contrast}
\title{Make color contrast}
\usage{
make_color_contrast(
  x,
  y,
  bg = "white",
  L_threshold = 60,
  C_floor = 130,
  L_hi = 90,
  L_lo = 40,
  do_plot = FALSE,
  cex = 2,
  ...
)
}
\description{
Make color contrast
}
\examples{
x <- c("red", "blue", "gold", "ivory");
y <- rep("red4", 4);
make_color_contrast(x, y);
make_color_contrast(x, y, C_floor=200);

y <- rep("gold", 4);
make_color_contrast(x, y);

y <- rep("darkorange", 4);
make_color_contrast(x, y, L_threshold=60);

y <- c("red4", "aquamarine4", "blue3", "yellow")
make_color_contrast(x, y, C_floor=150);

}
