% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-degree-groups.R
\name{make_degrees_clockwise}
\alias{make_degrees_clockwise}
\title{Make degrees angles into clockwise arc}
\usage{
make_degrees_clockwise(x)
}
\arguments{
\item{x}{\code{numeric} vector of degree angles.}
}
\value{
\code{data.frame} with three columns:
\itemize{
\item \code{'x'}: the \code{numeric} vector of degrees, sorted in increasing order,
with values adjusted so that the difference between angle
is positive, and has consistent numeric difference. When
the arc crosses zero, the first angle will therefore be negative.
\item \code{'idx'}: the \code{integer} index position of input \code{degrees},
which may be used as the equivalent of \code{order()}.
\item \code{'xdiff'}: the numeric difference between each degree angle,
and the next angle in the sequence. The last value should always
be the largest, representing the gap between the end of the arc,
and the start of the arc.
}
}
\description{
Make degrees angles into clockwise arc
}
\details{
The purpose is very specific, it takes a vector of degree angles,
determines the correct sequence of angles to constitute the
proper arc. A proper arc is defined as a series of points where
there is only one "largest gap" and this gap represents the
open space. When the arc crosses zero, angles are shifted
to negative values so that the numeric difference between
each angle is consistent.

For this purpose, "clockwise" is defined "increasing numeric order".

The assumption is that values in \code{x} only represent one contiguous arc,
that does not loop around itself, and therefore that values in \code{x}
may be provided in any order. Values returned will be sorted to start
with the first angle in the series, so that the last angle in the
series is positive and less than 360. If the arc crosses zero, the
first value in the series will be negative.
}
\examples{
x <- c(355, 5, 10, 350, 15, 20)
make_degrees_clockwise(x)

# re-order the original angles
x[make_degrees_clockwise(x)$idx]

make_degrees_clockwise(c(100, 200))
make_degrees_clockwise(c(200, 100, 300) + 110)
}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{assign_degree_groups}()},
\code{\link{check_systemfonts_family}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{render_venndir_footnotes}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{subset_systemfonts}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
