% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-overlap.R
\name{make_venn_combn_df}
\alias{make_venn_combn_df}
\title{Make full Venn combination data.frame}
\usage{
make_venn_combn_df(x, include_zero = FALSE, sep = "&", ...)
}
\arguments{
\item{x}{either character vector of set names, or integer number
of sets.}

\item{include_zero}{logical indicating whether to include one
row with all zeros, in the event of counting factors where
some factor levels may not be present in any sets.}

\item{sep}{character separator used when combining set names into
a single label.}

\item{...}{additional arguments are ignored.}
}
\description{
Make full Venn combination data.frame
}
\details{
This function returns a `data.frame` with all possible
combinations of set overlaps for the number of sets
provided.
}
\examples{
make_venn_combn_df(3);
make_venn_combn_df(3, include_zero=TRUE);

make_venn_combn_df(letters[1:3]);

}
