% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{match_list}
\alias{match_list}
\title{Match list elements to another list}
\usage{
match_list(x, y, ...)
}
\description{
Match list elements to another list
}
\examples{
x <- list(a=LETTERS[1],
   b=LETTERS[1:2],
   c=LETTERS[2:4]);
x;
y <- list(
   a=LETTERS[1:2],
   b=LETTERS[2],
   c=LETTERS[2:4]);
y;
match_list(x, y)

}
