% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{minus_polygon_list}
\alias{minus_polygon_list}
\title{Subtract one or more polygons}
\usage{
minus_polygon_list(polygon_list, new_name = NULL, ...)
}
\arguments{
\item{polygon_list}{\code{list} object that contains one or more polygons.}

\item{new_name}{\code{character} string with optional new name for the
output polygon.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{list} of polygons
}
\description{
Subtract one or more polygons
}
\details{
This function takes a \code{list} of polygons and iteratively
calls \code{polyclip::polyclip(A, B, op="minus")} to produce a union
across one or more polygons, which otherwise only works with two
polygons.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3)
counts <- c(A=1, B=2, `A&B`=3, C=5, `B&C`=2, `A&C`=2, `A&B&C`=1)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)

circle_minus <- minus_polygon_list(polygon_list);
circle_colors <- colorjam::rainbowJam(length(polygon_list), alpha=0.5);
plot_polygon_list(polygon_list, col=circle_colors, main="minus")
plot_polygon_list(circle_minus, col="#FFDD0088", border="red", lwd=3, add=TRUE)

circle_minus2 <- minus_polygon_list(polygon_list[c(2,1,3)]);
plot_polygon_list(circle_minus2, col="#FFDD0088", border="blue", lwd=3, add=TRUE)
}
\seealso{
Other venndir polygons: 
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_polygon_buffer}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{nudge_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_coord_list}()}
}
\concept{venndir polygons}
