% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-label-outside-jp.R
\name{nearest_point_JamPolygon}
\alias{nearest_point_JamPolygon}
\title{Get the nearest polygon point to a reference point}
\usage{
nearest_point_JamPolygon(x, jp, ...)
}
\arguments{
\item{x}{\code{numeric} matrix with columns \code{"x"} and \code{"y"}}

\item{jp}{\code{JamPolygon}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} matrix with columns \code{"x"} and \code{"y"}
}
\description{
Get the nearest polygon point to a reference point
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon,ANY,ANY,ANY-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
