% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{nudge_JamPolygon}
\alias{nudge_JamPolygon}
\title{Nudge JamPolygon coordinates}
\usage{
nudge_JamPolygon(
  jp,
  nudge = NULL,
  rotate_degrees = 0,
  center = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{jp}{\code{JamPolygon} object}

\item{nudge}{\code{list} whose names match \code{names(jp)}, containing \code{numeric}
vector with names \code{"x"} and \code{"y"}. For example:
\code{nudge=list(polyname1=c(x=1, y=0))}}

\item{...}{additional arguments are ignored}
}
\description{
Nudge JamPolygon coordinates
}
\details{
Polygon coordinates within a \code{JamPolygon} object are nudged by name
or polygon number, such that all parts of each polygon are adjusted
together. For multi-part polygons, and/or polygons with internal holes,
all parts are moved the identical amount.
}
\examples{
DEdf <- data.frame(check.names=FALSE,
   name=c("D", "E"),
   x=I(list(
      c(-3, 3, 3, 0, -3),
      c(-4, 2, 2, -4))),
   y=I(list(
      c(-3, -3, 1.5, 4, 1.5),
      c(-2, -2, 4, 4))),
   fill=c("#FFD70055", "#B2222255"))
DEjp <- new("JamPolygon", polygons=DEdf)
plot(DEjp)
nudge <- list(D=c(7, 1), E=c(-1, -1));
DEjp_nudged <- nudge_JamPolygon(DEjp, nudge=nudge)
plot(DEjp_nudged)

plot(rbind2(DEjp, DEjp_nudged),
   fill=c("#FFD70055", "#B2222255", "gold", "firebrick"),
   label=c("D_old", "E_old", "D_new", "E_new"),
   border.lty=c(2, 2, 1, 1))

}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{[,JamPolygon-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
