% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{nudge_polygon_list}
\alias{nudge_polygon_list}
\title{Nudge polygon_list}
\usage{
nudge_polygon_list(polygon_list = NULL, nudge = NULL, ...)
}
\arguments{
\item{polygon_list}{\code{list} object with \code{"x"} and \code{"y"} elements.}

\item{nudge}{\code{list} whose names are found in \code{names(polygon_list)},
and whose values are \code{x} and \code{y} coordinates to be moved.}

\item{...}{additional arguments are ignored.}

\item{rotate_degrees}{\code{numeric} value in degrees (0, 360) to
rotate the \code{polygon_list} object and all contained polygons.
(Not yet implemented.)}
}
\value{
object \code{list} polygon_list object with \code{"x"} and \code{"y"} elements.
}
\description{
Nudge polygon_list
}
\details{
This helper function is intended to take \code{list} polygon_list coordinates
and "nudge" (move by adding a scalar value to each coordinate)
only a subset of polygons identified by name.
}
\examples{
D <- list(
   x=c(-3, 3, 3, 0, -3),
   y=c(-3, -3, 1.5, 4, 1.5))
E <- list(
   x=c(-3, 3, 3, -3),
   y=c(-3, -3, 3, 3))
DElist <- list(D=D, E=E, DE=list(D=D, E=E))
nudge <- list(D=c(x=0, y=10), E=c(x=0, y=-10), DE=c(x=10, y=0))
new_polygon_list <- nudge_polygon_list(polygon_list=DElist,
   nudge=nudge)
poly_colors <- colorjam::rainbowJam(3, alpha=0.5);
plot_polygon_list(DElist, col=poly_colors)
plot_polygon_list(new_polygon_list, col=poly_colors)

polygon_list <- polygon_ellipses(c(3, 2), c(2, 3),
   xradius=c(1, 4),
   yradius=c(5, 2))
plot_polygon_list(polygon_list,
   col=c("#FF000077", "#FFDD0000"),
   xlim=c(-2, 9));
polygon_list2 <- nudge_polygon_list(polygon_list,
   nudge=list(`2`=c(x=3, y=-2))
)
plot_polygon_list(polygon_list2,
   col=c("#FF000077", "#FFDD0077"),
   add=TRUE,
   xlim=c(-2, 9));

plot_polygon_list(polygon_list[2], border="blue", lty="dotted", lwd=3, add=TRUE);
plot_polygon_list(polygon_list2[2], border="blue", lty="dotted", lwd=3, add=TRUE);
arrows(x0=2, x1=5, y0=3, y1=1)

}
\seealso{
Other venndir polygons: 
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_polygon_buffer}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_coord_list}()}
}
\concept{venndir polygons}
