% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{nudge_sp}
\alias{nudge_sp}
\title{Nudge SpatialPolygons}
\usage{
nudge_sp(sp, x_nudge = NULL, y_nudge = NULL, rotate_degrees = 0, ...)
}
\description{
Nudge SpatialPolygons
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{get_venn_shapes}()},
\code{\link{make_venn_test}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()}
}
\concept{venndir utility}
