% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{plot_polygon_list}
\alias{plot_polygon_list}
\title{Plot polygon_list using base R}
\usage{
plot_polygon_list(
  polygon_list,
  col = NULL,
  border = "black",
  lwd = 1,
  add = FALSE,
  asp = 1,
  bty = "n",
  xaxt = "n",
  yaxt = "n",
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  rule = c("evenodd", "none"),
  ...
)
}
\description{
Plot polygon_list using base R
}
\examples{
counts <- c(A=1, B=2, `A&B`=3, C=5, `B&C`=2, `A&C`=2, `A&B&C`=1)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
plot_polygon_list(polygon_list,
   col=colorjam::rainbowJam(length(polygon_list), alpha=0.5))

polygon_list2 <- list(A=polygon_list$A, BC=polygon_list[c("B", "C")])
plot_polygon_list(polygon_list2,
   col=colorjam::rainbowJam(length(polygon_list2), alpha=0.5))

}
