% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polygon-areas.R
\name{polygon_areas}
\alias{polygon_areas}
\title{Polygon area for simple or list of polygons}
\usage{
polygon_areas(x, y = NULL, simplify = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{numeric} input in one of the following formats:
\itemize{
\item \code{numeric} vector of coordinates for single polygon input, which
also requires \code{y} is supplied as \code{numeric} vector with equal length.
\item \code{list} of \code{numeric} vectors representing multiple polygons,
which also requires \code{y} is supplied as equivalent \code{list}.
\item \code{list} of polygons, each polygon contains elements \code{"x"} and \code{"y"}.
\item \code{list} of polygons, each polygon contains a \code{list} of polygon
component parts which each contain elements \code{"x"} and \code{"y"}.
}}

\item{y}{\code{numeric} vector or \code{list} of numeric vectors, compatible
with \code{x}, or \code{NULL} when \code{x} contains both coordinates.}

\item{simplify}{\code{logical} indicating whether area should be summed
for each polygon, potentially containing nested component polygons.
\itemize{
\item \code{simplify=TRUE} returns \code{numeric} vector with one total area
per polygon.
\item \code{simplify=FALSE} returns a \code{list} of \code{numeric} areas, using nested
list to indicate component polygons.
\item Note that this step does not manipulate the polygons in any way,
for example it does not call union over component polygons, therefore
the component polygons may overlap.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} vector with polygon area for each individual polygon
in the input \code{x},\code{y}.
\itemize{
\item When \code{x} is a \code{list} that contains \code{"x"} and \code{"y"} elements, those
elements are used.
\item When \code{x} and \code{y} both contain a \code{list} of \code{numeric} vectors, each
vector is considered coordinates of a polygon, and the area is returned
for each polygon.
\item When \code{x} and \code{y} are \code{numeric} vectors, it is considered a single
polygon, and thus one area is returned.
}
}
\description{
Polygon area for simple or list of polygons
}
\examples{
D <- list(
   x=c(-3, 3, 3, 0, -3),
   y=c(-3, -3, 1.5, 4, 1.5))
polygon_areas(D)

E <- list(
   x=c(-3, 3, 3, -3),
   y=c(-3, -3, 3, 3))
polygon_areas(E)

DElist <- list(
   x=list(
      D=c(-3, 3, 3, 0, -3),
      E=c(-3, 3, 3, -3)),
   y=list(
      D=c(-3, -3, 1.5, 4, 1.5),
      E=c(-3, -3, 3, 3)))
polygon_areas(DElist)

# list of polygons
poly_list <- list(D=D, E=E)
polygon_areas(poly_list)

# list of nested polygons
polygon_areas(list(DE=poly_list, D=D, E=E))

polygon_areas(list(DE=poly_list, D=D, E=E), simplify=TRUE)

}
\seealso{
Other venndir polygons: 
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_polygon_buffer}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_list}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_coord_list}()}
}
\concept{venndir polygons}
