% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{polygon_circles}
\alias{polygon_circles}
\title{Make polygon_list circles}
\usage{
polygon_circles(xcenter, ycenter, setnames = NULL, radius = 1, n = 60, ...)
}
\arguments{
\item{xcenter, ycenter}{\code{numeric} vector that defines the x and y
coordinate position of the center of each circle.}

\item{setnames}{\code{vector} that contains names for each circle, stored
as \code{names()} for each polygon.
When \code{setnames} is \code{NULL} then index numbers are used.}

\item{radius}{\code{numeric} vector that defines the radius of each circle.
This \code{vector} is recycled to \code{length(xcenter)}.}

\item{n}{\code{integer} value indicating the number of subdivisions to
use in the circle.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{JamPolygon} object
}
\description{
Make polygon_list circles
}
\details{
This function creates one or more circles as polygon_list \code{list} objects.
}
\examples{
circle_jp <- polygon_circles(c(3, 2), c(2, 3))
plot(circle_jp, fill=c("red", "gold"))

}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
