% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{polygon_ellipses}
\alias{polygon_ellipses}
\title{Make polygon_list ellipses}
\usage{
polygon_ellipses(
  xcenter,
  ycenter,
  setnames = NULL,
  xradius = 1,
  yradius = 2,
  rotation_degrees = c(0),
  n = 60,
  ...
)
}
\arguments{
\item{xcenter, ycenter}{\code{numeric} vector that defines the x and y
coordinate position of the center of each ellipse.}

\item{setnames}{\code{vector} that contains names for each circle, stored
as \code{names()} for each polygon.
When \code{setnames} is \code{NULL} then index numbers are used.}

\item{xradius, yradius}{\code{numeric} vector that defines the radius
of each ellipse along the x-axis and y-axis, respectively.
Each \code{vector} is recycled to \code{length(xcenter)}.}

\item{n}{\code{integer} value indicating the number of subdivisions to
use in the circle.}

\item{...}{additional arguments are ignored.}

\item{rotation_degree}{\code{numeric} vector representing degrees to
rotate each ellipse after it is created, where values are
conformed to between \code{0} and \code{360}, rotating clockwise.}
}
\value{
object \code{list} with a number of circles encoded as polygons.
}
\description{
Make polygon_list ellipses
}
\details{
This function creates one or more ellipses as polygon_list \code{list} objects.
}
\examples{
polygon_list <- polygon_ellipses(c(3, 2), c(2, 3),
   xradius=c(1, 4),
   yradius=c(5, 2))
plot_polygon_list(polygon_list, col=c("#FF000077", "#FFDD0077"));
points(x=c(3, 2), y=c(2, 3), pch=c("1", "2"), add=TRUE);

}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
