% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{polygon_list_labelr}
\alias{polygon_list_labelr}
\title{Simple wrapper to polylabelr::poi() for polygon_list}
\usage{
polygon_list_labelr(polygon_list, precision = 1, add_labels = FALSE, ...)
}
\arguments{
\item{polygon_list}{\code{list} object}
}
\value{
\code{matrix} with nrow \code{length(polygon_list)} with x,y coordinates
representing the visual center of each polygon in the list.
}
\description{
Simple wrapper to polylabelr::poi() for polygon_list
}
\examples{
counts <- c(A=1, B=2, `A&B`=3, C=5, `B&C`=2, `A&C`=2, `A&B&C`=1)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)

# default is to label each polygon in its center
plot_polygon_list(polygon_list,
   col=colorjam::rainbowJam(length(polygon_list), alpha=0.5))
polygon_list_labelr(polygon_list, add_labels=TRUE)

# create unique polygons for each label
A_only <- minus_polygon_list(polygon_list, new_name="A_only");
B_only <- minus_polygon_list(polygon_list[c(2,1,3)], new_name="B_only");
C_only <- minus_polygon_list(polygon_list[c(3,1,2)], new_name="C_only");

plot_polygon_list(polygon_list,
   col=colorjam::rainbowJam(length(polygon_list), alpha=0.5))
ABC_only <- c(A_only, B_only, C_only);
polygon_list_labelr(ABC_only, add_labels=TRUE)

# label ABC intersection
ABC_int <- intersect_polygon_list(polygon_list[c(1,2,3)], new_name="ABC");
plot_polygon_list(ABC_int, add=TRUE, col="gold")
polygon_list_labelr(ABC_int, add_labels=TRUE)

# label AB intersection
AB_only <- minus_polygon_list(
   c(intersect_polygon_list(polygon_list[c(1,2)], new_name="BC_only"),
      polygon_list[3]))
plot_polygon_list(AB_only, add=TRUE, col="darkviolet")
polygon_list_labelr(AB_only, add_labels=TRUE, col="white")

# label BC intersection
BC_only <- minus_polygon_list(
   c(intersect_polygon_list(polygon_list[c(2,3)], new_name="BC_only"),
      polygon_list[1]))
plot_polygon_list(BC_only, add=TRUE, col="skyblue")
polygon_list_labelr(BC_only, add_labels=TRUE)

# test with fully overlapping polygon (to create a hole)
counts <- c(A=5, B=0, C=3, `A&B`=1)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
plot_polygon_list(polygon_list[1:3], col=c("red"))
A_only <- minus_polygon_list(polygon_list[c(1, 2, 3)], new_name="A_only");
plot_polygon_list(A_only, col="gold", add=TRUE)
polygon_list_labelr(A_only, add_labels=TRUE)

polygon_list_labelr(c(A_only, polygon_list[2:3]), add_labels=TRUE)
}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
