% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-poly-conversions.R
\name{polygon_list_to_xy_list}
\alias{polygon_list_to_xy_list}
\title{Convert polygon list of x,y coordinate into a list by x and y}
\usage{
polygon_list_to_xy_list(x, flatten = FALSE, ...)
}
\arguments{
\item{x}{\code{list} of polygons
\itemize{
\item each polygon should contain a \code{list} with elements \code{"x"} and \code{"y"}.
\item each polygon can contain multiple component polygons as a
nested list, in which case this function is called iteratively
so that the component \code{"x"} and \code{"y"} are returned as equivalent
nested \code{list} objects.
\item In all cases, \code{names(output$x)} and \code{names(output$y)} should equal
\code{names(x)}.
}}

\item{flatten}{\code{logical} indicating whether all polygons should be
flattened to the same level, without nested polygons.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} with elements \code{"x"} and \code{"y"} which each contain a
\code{list} with length \code{length(x)}.
}
\description{
Convert polygon list of x,y coordinate into a list by x and y
}
\details{
Input is a list of polygons, where each polygon contains a \code{list}
with elements \code{"x"} and \code{"y"}. Output is a list of \code{"x"} and \code{"y"}
split by each polygon.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
jamba::ssdim(polygon_list)
jamba::ssdim(polygon_list_to_xy_list(polygon_list))

jamba::ssdim(polygon_list_to_xy_list(list(AB=polygon_list)))

}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
