% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-printcolordf.R
\name{print_color_df}
\alias{print_color_df}
\title{Print colorized text table}
\usage{
print_color_df(
  df,
  dfcolor = NULL,
  dfinvert = NULL,
  colnames = FALSE,
  padding = 2,
  ...
)
}
\description{
Print colorized text table
}
