% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-footnote.R
\name{render_venndir_footnotes}
\alias{render_venndir_footnotes}
\title{Add Venndir footnotes to a Venndir figure}
\usage{
render_venndir_footnotes(
  venndir_output,
  x = "bottomright",
  footnote_x_inset = grid::unit(0, "lines"),
  footnote_y_inset = grid::unit(0.5, "lines"),
  footnote_fontfamily = NULL,
  footnote_color = "black",
  footnote_fontsize = 14,
  footnote_outline = TRUE,
  footnote_style = c("symbols", "header", "footnote"),
  draw_footnote = TRUE,
  use_devoid = getOption("use_devoid", TRUE),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{venndir_output}{\code{Venndir} object}

\item{x}{\code{character} position for the footnote}

\item{footnote_fontfamily, footnote_color, footnote_fontsize}{arguments
to customize the font and color used to render the footnote.}

\item{footnote_outline}{\code{logical} whether to draw an outline around the
footnote for visible clarity, default TRUE.
It calls \code{jamba::setTextContrastColor()} using \code{footnote_color}.}

\item{draw_footnote}{\code{logical}, default TRUE, whether to draw the resulting
footnote when one or more footnotes are defined.
When \code{FALSE} a grid \code{grob} is returned if footnotes exist, otherwise
it returns NULL.}

\item{...}{additional arguments are ignored.}
}
\description{
Add Venndir footnotes to a Venndir figure
}
\details{
This function adds a subtle footnote symbol, or symbols, if
they are defined in the 'metadata' slot of the input \code{Venndir}
object, entry 'footnotes'. Footnote symbols are comma-delimited
and displayed together.

The intention is to have a visibly distinct indication when some
footnote exists and should be reviewed.
}
\examples{
setlist3 <- make_venn_test(500, n_sets=4)
v3 <- venndir(setlist3,
   shape="ellipse",
   vector_method="label",
   proportional=TRUE)
footnotes(v3)
render_venndir_footnotes(v3)

v3 <- venndir(setlist3,
   shape="ellipse",
   footnote_symbols="\u203B",
   proportional=TRUE)
footnotes(v3)

}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{assign_degree_groups}()},
\code{\link{check_systemfonts_family}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_degrees_clockwise}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{subset_systemfonts}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
