% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-reposition-labels.R
\name{reposition_venn_gridtext_labels}
\alias{reposition_venn_gridtext_labels}
\title{Reposition venn gridtext labels}
\usage{
reposition_venn_gridtext_labels(
  g_labels,
  gdf,
  segment_df = NULL,
  do_draw = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g_labels}{\code{richtext_grob} class as returned by
\code{gridtext::richtext_grob()}.}

\item{gdf}{\code{data.frame} with one row per \code{grob} in \code{g_labels},
with colnames \code{"x"}, \code{"y"}, \code{"overlap_set"}, \code{"r"},
\code{"r_unit"}, \code{"border_col"}, \code{"box_fill"}. The rows
are grouped using \code{c("x", "y", "overlap_set")}.
The \code{"r"} and \code{"r_unit"} values are used to define
the radius and radius unit for the rounded corner.
The row in \code{gdf} with the highest \code{"r"} value is the
reference row, for the \code{"r"} radius, and the \code{"border_col"}
and \code{"box_fill"} values.}

\item{segment_df}{\code{data.frame} with segment coordinates, optional.}

\item{do_draw}{\code{logical} indicating whether to draw the finished
result in the context of the current open graphics device.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Reposition venn gridtext labels
}
\details{
This is an internal function, based upon \code{draw_gridtext_groups()}
which takes grid grobs already defined in context of an active
graphics device, so the actual height and width is defined,
and repositions labels for each group of labels.

For example, when the group of labels is connected to a line
segment on the right side, the labels will be right-justified
so the line segment ends at the right edge of the labels.

Also, this function is intended to allow different "layouts",
such as having the set title at top-center, with counts in
two columns below the title on the left and right.
}
\seealso{
Other venndir label: 
\code{\link{draw_gridtext_groups}()},
\code{\link{polygon_label_fill}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{polygon_list_label_segment}()},
\code{\link{sp_polylabelr}()}
}
\concept{venndir label}
