% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{rescale_coordinates}
\alias{rescale_coordinates}
\title{Scale, rotate, and shift numeric coordinates}
\usage{
rescale_coordinates(
  x,
  scale = c(1, 1),
  rotate_degrees = 0,
  shift = c(0, 0),
  center = NULL,
  rotation_axes = c(1, 2),
  plot_debug = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{matrix} with one or more columns containing \code{numeric}
coordinates.}

\item{scale}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. After subtracting the \code{center}, the coordinates
in each column are multiplied by the \code{scale}.}

\item{rotate_degrees}{\code{numeric} value in degrees indicating
rotation around the \code{center}, where positive values are
clockwise rotation. This rotation is only applied to two
columns in \code{x} defined by \code{rotation_axes}.}

\item{shift}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. The coordinates in each column are added to
the \code{shift}, after applying \code{scale} and \code{rotate_degrees}}

\item{center}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}, indicating the center point used for \code{scale} and
\code{rotate_degrees} transformations. When \code{center=NULL} it
is derived from the bounding box, which is the mean of the range
for each column in \code{x}.}

\item{rotation_axes}{\code{integer} vector length 2, indicating which
two columns in \code{x} are used for \code{rotate_degrees}.}

\item{plot_debug}{\code{logical} indicating whether to plot starting,
intermediate, and final polygons during the processing.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{matrix} with \code{numeric} values after processing.
}
\description{
Scale, rotate, and shift numeric coordinates
}
\details{
This function takes a numeric matrix with two or more
numeric columns, and adjusts the coordinates in three
ways:
\itemize{
\item scale: adjust coordinate range by a multiplier, relative to
a central point
\item rotate: rotate coordinates around a central point in degrees
\item shift: adjust coordinate range by adding a numeric value
}

The operations are performed in that order: rotate, scale, shift.

When \code{center} is not defined, the default behavior is to use
the mean of the range of each coordinate column. Using the mean
range is equivalent to using the mean of the bounding box \code{sp::bbox()}
for Spatial objects.
}
\examples{
sp <- sp_ellipses(xcenter=2, ycenter=3, xradius=3, yradius=1);

plot(sp, col="#FF000077", border="#FF0000");
axis(1, las=2); axis(2, las=2);

x <- sp@polygons[[1]]@Polygons[[1]]@coords;

jamba::nullPlot(doBoxes=FALSE, xlim=c(-3, 9), ylim=c(-2, 10), asp=1);
xnew <- rescale_coordinates(x,
   shift=c(1, 1),
   rotate_degrees=30,
   scale=c(1.5, 2.5),
   plot_debug=TRUE)
axis(1, las=2); axis(2, las=2);
   
# example using SpatialPolygons
sp <- sp_ellipses(xcenter=c(2, 4),
   ycenter=c(3, 2),
   xradius=c(3, 2),
   yradius=c(1, 3));
# apply rescale to each polygon
sp_2 <- rescale_sp(sp,
   shift=c(1, 1),
   rotate_degrees=80,
   scale=c(1.5, 2.5))
par("mfrow"=c(1,2));
plot(sp,
   asp=1,
   col=jamba::alpha2col(alpha=0.7, c("red", "gold")),
   main="Input ellipses");
plot(sp_2,
   asp=1,
   col=jamba::alpha2col(alpha=0.7, c("red", "gold")),
   main="rescaled ellipses");
par("mfrow"=c(1,1));

# same as before but apply 180 degree rotation
sp_3 <- rescale_sp(sp,
   shift=c(1, 1),
   rotate_degrees=160,
   scale=c(1, 1),
   share_center=TRUE)
par("mfrow"=c(1,2));
plot(sp, asp=1,
   col=jamba::alpha2col(alpha=0.7, c("red", "gold")));
plot(sp_3, asp=1,
   col=jamba::alpha2col(alpha=0.7, c("red", "gold")));
par("mfrow"=c(1,1));

}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{intersect_polygons}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{rescale_sp}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()}
}
\concept{venndir spatial}
