% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{rescale_p}
\alias{rescale_p}
\title{Rescale a Polygon object}
\usage{
rescale_p(
  p,
  rotate_degrees = 0,
  scale = c(1, 1),
  shift = c(0, 0),
  center = NULL,
  ...
)
}
\arguments{
\item{p}{object \code{sp::Polygon}}

\item{rotate_degrees}{\code{numeric} value in degrees indicating
rotation around the \code{center}, where positive values are
clockwise rotation. This rotation is only applied to two
columns in \code{x} defined by \code{rotation_axes}.}

\item{scale}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. After subtracting the \code{center}, the coordinates
in each column are multiplied by the \code{scale}.}

\item{shift}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. The coordinates in each column are added to
the \code{shift}, after applying \code{scale} and \code{rotate_degrees}}

\item{center}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}, indicating the center point used for \code{scale} and
\code{rotate_degrees} transformations. When \code{center=NULL} it
is derived from the bounding box, which is the mean of the range
for each column in \code{x}.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{sp::Polygon}
}
\description{
Rescale a Polygon object
}
\details{
This function simply applies \code{rescale_coordinates()} to an
object \code{sp::Polygon}, and it does so by calling
\code{rescale_coordinates} on each \code{coords} object contained
in the \code{p} input.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{gridtext_make_outer_box}()},
\code{\link{gridtext_richtext_grob}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_venndir_label}()},
\code{\link{print_color_df}()},
\code{\link{rescale_ps}()},
\code{\link{shrink_df}()},
\code{\link{sp_percent_area}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_legender}()}
}
\concept{venndir utility}
