% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{rescale_polygon_list}
\alias{rescale_polygon_list}
\title{Rescale a polygon_list object}
\usage{
rescale_polygon_list(
  polygon_list,
  rotate_degrees = 0,
  scale = c(1, 1),
  shift = c(0, 0),
  center = NULL,
  share_center = FALSE,
  share_polygon_center = TRUE,
  ...
)
}
\arguments{
\item{polygon_list}{\code{list} object}

\item{rotate_degrees}{\code{numeric} value in degrees indicating
rotation around the \code{center}, where positive values are
clockwise rotation. This rotation is only applied to two
columns in \code{x} defined by \code{rotation_axes}.}

\item{scale}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. After subtracting the \code{center}, the coordinates
in each column are multiplied by the \code{scale}.}

\item{shift}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. The coordinates in each column are added to
the \code{shift}, after applying \code{scale} and \code{rotate_degrees}}

\item{center}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}, indicating the center point used for \code{scale} and
\code{rotate_degrees} transformations. When \code{center=NULL} it
is derived from the bounding box, which is the mean of the range
for each column in \code{x}.}

\item{share_center}{\code{logical} indicating whether all polygons
should share the same center, where \code{share_center=TRUE} will
adjust everything collectively, and \code{share_center=FALSE} will
adjust each polygon independently relative to its own center
coordinate.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{list} polygon_list
}
\description{
Rescale a polygon_list object
}
\details{
This function simply applies \code{rescale_coordinates()} to an
\code{list} polygon_list object.
}
\examples{
polygon_list <- polygon_ellipses(c(3, 2), c(2, 3),
   xradius=c(1, 4),
   yradius=c(5, 2))
polygon_list1 <- intersect_polygon_list(polygon_list);
polygon_list2 <- minus_polygon_list(polygon_list[1:2]);
polygon_list3 <- minus_polygon_list(polygon_list[2:1]);
polygon_list123 <- c(polygon_list1,
   polygon_list2,
   polygon_list3);

polygon_list123a <- rescale_polygon_list(polygon_list123,
   scale=c(1.5, 1.5),
   share_center=TRUE);
polygon_list123b <- rescale_polygon_list(polygon_list123,
   scale=c(1.5, 1.5));
col3 <- c("#FF000077", "#FFDD0077", "#0000DD77");
par("mfrow"=c(2, 2));
plot_polygon_list(polygon_list123,
   col=col3,
   main="original polygons",
   xlim=c(-10, 15), ylim=c(-5, 10));
axis(1, las=2); axis(2, las=2);
plot_polygon_list(polygon_list123a,
   col=col3,
   main="share_center=TRUE",
   xlim=c(-10, 15), ylim=c(-5, 10));
axis(1, las=2); axis(2, las=2);
plot_polygon_list(polygon_list123[1:2],
   col=col3[1:2],
   main="share_center=FALSE\nrescaling only the blue polygon",
   xlim=c(-10, 15), ylim=c(-5, 10));
axis(1, las=2); axis(2, las=2);
plot_polygon_list(polygon_list123b[3],
   col=col3[3],
   add=TRUE);
plot_polygon_list(polygon_list123[2:3],
   col=col3[2:3],
   main="share_center=FALSE\nrescaling only the red polygon",
   xlim=c(-10, 15), ylim=c(-5, 10));
axis(1, las=2); axis(2, las=2);
plot_polygon_list(polygon_list123b[1],
   col=col3[1],
   add=TRUE);
par("mfrow"=c(1, 1));

{par("mfrow"=c(2, 2));
plot_polygon_list(polygon_list123, col=col3,
   xlim=c(-4, 8), ylim=c(-4, 8))
title(main="Original polygons", line=0);
plot_polygon_list(rescale_polygon_list(polygon_list123, rotate_degrees=c(`11`=45, `12`=-10)), col=col3,
   xlim=c(-4, 8), ylim=c(-4, 8))
title(sub="yellow +45 degrees\nblue -10 degrees", line=0,
   main="share_polygon_center=TRUE (default)")
plot_polygon_list(rescale_polygon_list(polygon_list123, rotate_degrees=c(`11`=45, `12`=-10), share_polygon_center=FALSE), col=col3,
   xlim=c(-4, 8), ylim=c(-4, 8))
title(sub="yellow +45 degrees\nblue -10 degrees", line=0,
   main="share_polygon_center=FALSE\n(each polygon uses its center)")
plot_polygon_list(rescale_polygon_list(polygon_list123, rotate_degrees=c(`11`=45, `12`=-10), share_center=TRUE), col=col3,
   xlim=c(-4, 8), ylim=c(-4, 8))
title(sub="yellow +45 degrees\nblue -10 degrees", line=0,
   main="share_center=TRUE\n(all polygons share one global center)")
par("mfrow"=c(1, 1));}


}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
