% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{simple_ellipse}
\alias{simple_ellipse}
\title{Simple ellipse function}
\usage{
simple_ellipse(h, k, a, b = a, phi = 0, n = 200L)
}
\description{
Simple ellipse function
}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
