% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{sp_circles}
\alias{sp_circles}
\title{Make SpatialPolygons circles}
\usage{
sp_circles(xcenter, ycenter, setnames = NULL, radius = 1, n = 60, ...)
}
\arguments{
\item{xcenter, ycenter}{\code{numeric} vector that defines the x and y
coordinate position of the center of each circle.}

\item{setnames}{\code{vector} that contains names for each circle, stored
in the \code{sp::SpatialPolygons} object as the polygon name. When
\code{setnames} is \code{NULL} then integer index positions are used.}

\item{radius}{\code{numeric} vector that defines the radius of each circle.
This \code{vector} is recycled to \code{length(xcenter)}.}

\item{n}{\code{integer} value indicating the number of subdivisions to
use in the circle.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{sp::SpatialPolygons} with a number of circles
defined by \code{length(xcenter)}.
}
\description{
Make SpatialPolygons circles
}
\details{
This function creates one or more circles as
\code{sp::SpatialPolygons} objects.
}
\examples{
sp <- sp_circles(c(3, 2), c(2, 3))
plot(sp, col=c("#FF000077", "#FFDD0077"));
axis(1, las=2); axis(2, las=2);
points(x=c(3, 2), y=c(2, 3), pch=c("1", "2"), add=TRUE);

}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{intersect_polygons}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_sp}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()}
}
\concept{venndir spatial}
