% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{sp_ellipses}
\alias{sp_ellipses}
\title{Make SpatialPolygons ellipses}
\usage{
sp_ellipses(
  xcenter,
  ycenter,
  setnames = NULL,
  xradius = 1,
  yradius = 2,
  rotation_degrees = c(0),
  n = 60,
  ...
)
}
\description{
Make SpatialPolygons ellipses
}
