% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{sp_polylabelr}
\alias{sp_polylabelr}
\title{Get SpatialPolygons polylabelr coordinate}
\usage{
sp_polylabelr(sp, apply_holes = TRUE, ...)
}
\arguments{
\item{sp}{\code{sp::SpatialPolygons} object}

\item{apply_holes}{\code{logical} indicating whether to apply any
polygon holes when present in \code{sp}.}

\item{...}{additional arguments are ignored}
}
\value{
\code{list} with items \code{x} with x coordinate, \code{y} with
y coordinate, and \code{dist} with distance to the enclosing
polygon.
}
\description{
Get SpatialPolygons polylabelr coordinate
}
\details{
This function is a simple wrapper function around
\code{polylabelr::poi()} for \code{sp::SpatialPolygons} input.
This function does two things:
\enumerate{
\item It finds the largest polygon in the \code{sp::SpatialPolygons}
input, which may contain multiple disconnected polygons.
\item It applies "holes" inside the polygon when present,
to prevent the label from being chosen inside the hole.
}
}
\examples{
setlist <- list(A=letters, B=sample(letters, 4));
vo <- venndir(setlist, proportional=TRUE, do_plot=FALSE);
sp <- vo$venn_spdf[1,];
plot(sp, col="gold")
xy <- sp_polylabelr(sp);
points(xy, pch=20, cex=4, col="navy")

# example with a hole inside
sp <- sp_ellipses(xcenter=1, ycenter=1, xradius=0.5, yradius=1);
sp_hole <- sp_ellipses(xcenter=1, ycenter=1, xradius=0.2, yradius=0.5);
sp_donut <- rgeos::gDifference(sp, sp_hole);
plot(sp_donut, col="gold")
spt <- sp_polylabelr(sp_donut);
text(x=spt$x, y=spt$y, add=TRUE, labels="label")

}
\seealso{
Other venndir label: 
\code{\link{draw_gridtext_groups}()},
\code{\link{polygon_label_fill}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{reposition_venn_gridtext_labels}()}
}
\concept{venndir label}
