% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{split_JamPolygon}
\alias{split_JamPolygon}
\title{Split JamPolygon multipart polygons}
\usage{
split_JamPolygon(jp, suffix = "_sub", ...)
}
\arguments{
\item{jp}{\code{JamPolygon}}

\item{suffix}{\code{character} passed to \code{jamba::makeNames()} when creating
new unique \code{names()} for the resulting polygons.}

\item{...}{additional arguments are ignored.}
}
\description{
Split JamPolygon multipart polygons
}
\seealso{
Other JamPolygon: 
\code{\link{area_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
