% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-systemfonts.R
\name{subset_systemfonts}
\alias{subset_systemfonts}
\title{Subset system fonts for review}
\usage{
subset_systemfonts(
  ...,
  fonts_df = systemfonts::system_fonts(),
  do_plot = FALSE,
  return_type = c("data.frame"),
  trim_path = TRUE,
  do_sort = TRUE,
  byCols = c("family", "italic", "weight", "style", "width", "name"),
  remove_problem_fonts = TRUE
)
}
\arguments{
\item{...}{additional arguments are passed to \code{subset()} to subset
the resulting table accordingly. This argument is first to
avoid matching other named arguments below.}

\item{fonts_df}{\code{data.frame} full set of fonts, default uses
\code{systemfonts::system_fonts()} to include all fonts recognized
by systemfonts.}

\item{do_plot}{\code{logical} whether to plot the search results using marquee,
default FALSE.}

\item{return_type}{\code{character} string with return class, default is
'data.frame' which converts the tibble output to data.frame
for convenience.}

\item{trim_path}{\code{logical} whether to trim the font file path, default
TRUE removes the folder by calling \code{basename()}.}

\item{do_sort}{\code{logical} whether to sort output, using 'byCols'.}

\item{byCols}{\code{character} columns used when \code{do_sort==TRUE}.}

\item{remove_problem_fonts}{\code{logical} default TRUE, whether to check for
errors accessing each font family by calling \code{systemfonts::font_info()}.}
}
\value{
\code{data.frame} with font information derived from
\code{systemfonts::system_fonts()} and subset according to '...'.
}
\description{
Subset system fonts for review, to organize fonts available to systemfonts
}
\details{
This function provides a convenient method to subset fonts
recognized by \code{systemfonts::system_fonts()}, while also sorting
and filtering out "problem fonts". The argument \code{do_plot=TRUE}
will visualize the resulting fonts using \code{marquee::marquee_grob()}.

Note that there are some "problem fonts" which are not recognized
by the Freetype API used by systemfonts, which appear to arise
from newer font file formats released on MacOS which are not yet
recognized by Freetype. Using one of these fonts may cause R crash,
therefore some steps were taken to avoid this possibility:
\itemize{
\item The output will include one new column \code{"check"} with integer number
of valid associated font files, usually \code{1} for valid, and \code{0} for no
valid files.
\item The argument \code{remove_problem_fonts=TRUE} removes any fonts with \code{0},
thereby helping to avoid crashing the R session.
\item The function \code{check_systemfonts_family()} will check one or more
font family values, returning the integer number of valid font files.
}
}
\examples{
subset_systemfonts(grepl("(Arial(.*Narrow|)|Helvetica)$", family))
subset_systemfonts(grepl("(Arial(.*Narrow|)|Helvetica.*)$", family), do_plot=TRUE)
subset_systemfonts(grepl("Helvetica Neue", family), do_plot=FALSE)

subset_systemfonts(grepl("serif", ignore.case=TRUE, name), do_plot=TRUE)

}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{assign_degree_groups}()},
\code{\link{check_systemfonts_family}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_degrees_clockwise}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{render_venndir_footnotes}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
