% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-textvenn.R
\name{textvenn}
\alias{textvenn}
\title{Text Venn diagram}
\usage{
textvenn(
  setlist,
  sets = seq_along(setlist),
  set_colors = NULL,
  spacing = 5,
  padding = 2,
  useBgColor = TRUE,
  countsUseBgColor = FALSE,
  return_items = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Text Venn diagram
}
\details{
This function is a very simple method to print a Venn diagram
using text, intended to be displayed using mono-spaced font
on an R console.
}
