% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{three_point_angle}
\alias{three_point_angle}
\title{Calculate angle between three consecutive points}
\usage{
three_point_angle(x)
}
\arguments{
\item{x}{\code{matrix} with first two columns assumed to contain \code{x}, \code{y}
coordinates.}
}
\description{
Calculate angle between three consecutive points
}
