% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-plotly.R
\name{to_basic.GeomTextBox}
\alias{to_basic.GeomTextBox}
\title{Basic function for GeomTextBox compatibility with ggplotly}
\usage{
to_basic.GeomTextBox(data, prestats_data, layout, params, p, ...)
}
\arguments{
\item{data, prestats_data, layout, params, p}{arguments used
by the ggplot2 infrastructure.}

\item{...}{additional arguments are passed along}
}
\description{
Basic function for GeomTextBox compatibility with ggplotly
}
\details{
Note that this function is still being tested, but is intended
to be used with ggplot2 output from \code{render_venndir()} so
it can be used with \code{plotly::ggplotly()}.
}
\seealso{
Other venndir plotly: 
\code{\link{to_basic.GeomRichText}()}
}
\concept{venndir plotly}
