% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{union_polygon_list}
\alias{union_polygon_list}
\title{Union one or more polygons}
\usage{
union_polygon_list(polygon_list, ...)
}
\arguments{
\item{polygon_list}{\code{list} object that contains one or more polygons.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{list} of polygons
}
\description{
Union one or more polygons
}
\details{
This function takes a \code{list} of polygons and iteratively
calls \code{polyclip::polyclip(A, B, op="union")} to produce a union
across one or more polygons, which otherwise only works with two
polygons.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)

circle_union <- union_polygon_list(polygon_list);
jamba::ssdim(circle_union)
circle_colors <- colorjam::rainbowJam(2);
plot_polygon_list(polygon_list, col=circle_colors, main="union")
plot_polygon_list(circle_union, col="#FFDD0088", border="red", lwd=3, add=TRUE)

counts2 <- c(A=1, B=2, `A&B`=3, C=4)
x2 <- eulerr::euler(counts2)
polygon_list2 <- eulerr_to_polygon_list(x2)
plot_polygon_list(polygon_list2)

}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
