% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{union_polygons}
\alias{union_polygons}
\title{Union one or more polygons}
\usage{
union_polygons(sp, ...)
}
\arguments{
\item{sp}{\code{list} object that contains one or more
\code{sp::SpatialPolygons}.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{sp::SpatialPolygons}
}
\description{
Union one or more polygons
}
\details{
This function takes a \code{list} of \code{sp::SpatialPolygons}
objects, and performs iterative \code{rgeos::gUnion()} on
the full list, since \code{rgeos::gUnion()} only operates on
two polygons at a time.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_percent_area}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
