% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{update_JamPolygon}
\alias{update_JamPolygon}
\title{Update attributes for a JamPolygon object}
\usage{
update_JamPolygon(jp, ...)
}
\arguments{
\item{...}{additional arguments are ignored.}
}
\description{
Update attributes for a JamPolygon object, including default label
position, and polygon orientations.
}
\details{
The default label position is defined with \verb{"label_x","label_y"}
if these columns do not already exist in \code{jp@polygons}.
The coordinates are defined by \code{labelr_JamPolygon()} with
argument \code{add_to_jp=TRUE}.

If the column \code{"orientation"} does not exist in \code{jp@polygons}
it is added by calling \code{add_orientation_JamPolygon()}. It
subsequently adds columns \code{"holes"}, \code{"polygon_clockwise"}, and
\code{"polygon_parent"} when each is not already present.
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{[,JamPolygon-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()}
}
\concept{JamPolygon}
