% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndir-class.R
\docType{methods}
\name{plot,Venndir,ANY-method}
\alias{plot,Venndir,ANY-method}
\alias{length,Venndir-method}
\alias{setlist,Venndir-method}
\alias{names,Venndir-method}
\alias{show,Venndir-method}
\alias{metadata,Venndir-method}
\alias{metadata<-,Venndir-method}
\alias{overlaplist,Venndir-method}
\alias{overlapdf,Venndir-method}
\alias{signed_counts,Venndir-method}
\alias{im,Venndir-method}
\title{Plot Venndir object}
\usage{
\S4method{plot}{Venndir,ANY}(x, y, ...)

\S4method{length}{Venndir}(x)

\S4method{setlist}{Venndir}(x)

\S4method{names}{Venndir}(x)

\S4method{show}{Venndir}(object)

\S4method{metadata}{Venndir}(x)

\S4method{metadata}{Venndir}(x) <- value

\S4method{overlaplist}{Venndir}(x)

\S4method{overlapdf}{Venndir}(x)

\S4method{signed_counts}{Venndir}(x)

\S4method{im}{Venndir}(x)
}
\value{
\code{Venndir} object, invisibly.
}
\description{
Plot Venndir object
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{expand_range}()},
\code{\link{get_venndir_curate_df}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{modify_venndir_overlap}()},
\code{\link{nudge_venndir_label}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{three_point_angle}()},
\code{\link{venndir_legender}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir utility}
