% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir.R
\name{venndir}
\alias{venndir}
\title{Directional Venn diagram}
\usage{
venndir(
  setlist,
  sets = seq_along(setlist),
  set_colors = NULL,
  overlap_type = c("concordance", "each", "overlap", "concordant"),
  proportional = FALSE,
  return_items = FALSE,
  display_zero = TRUE,
  font_cex = c(1, 0.8),
  poly_alpha = 0.5,
  label_style = c("basic", "fill", "shaded", "shaded_box", "lite", "lite_box"),
  x_nudge = NULL,
  y_nudge = NULL,
  unicode = TRUE,
  ...
)
}
\arguments{
\item{setlist}{`list` of named vectors, whose names represent
set items, and whose values represent direction using 
values `c(-1, 0, 1)`.}

\item{sets}{`integer` index of sets in `setlist` to display in
the Venn diagram. This subset is useful when creating a Venn
diagram for a subset of a list, because it defines consistent
colors across all sets, and uses the appropriate subset of
colors in the Venn diagram.}

\item{set_colors}{`character` vector of R colors, or `NULL` to
use default colors defined by `colorjam::rainbowJam()`.}

\item{overlap_type}{`character` value indicating the type of
overlap logic: `"each"` records each combination of
signs; `"overlap"` disregards the sign and returns any match
item overlap; `"concordance"` represents counts for full
agreement, or `"mixed"` for any inconsistent overlapping
direction; `"agreement"` represents full agreement in direction
as `"agreement"`, and `"mixed"` for any inconsistent
direction.}

\item{proportional}{`logical` indicating whether the Venn circles
are proportionally sized, also known as a Euler diagram. Note
that proportionally sized circles are not guaranteed to represent
every possible overlap.}

\item{return_items}{`logical` indicating whether to return
the items within each overlap set.}

\item{display_zero}{`logical` indicating whether empty overlaps
are labeled with zero `0` when `display_zero=TRUE`, or are
blank when `display_zero=FALSE`.}

\item{font_cex}{`numeric` `vector` length 2, indicating relative
font size for the main label, and directional label, respectively.
The default `c(1, 0.8)` defines the main label with 1x size,
and the directional labels with 80% that size.}

\item{poly_alpha}{`numeric` value between 0 and 1, indicating the
alpha transparency of the polygon backgroun color, where
`poly_alpha=1` is completely 100% opaque (no transparency), and
`poly_alpha=0.4` is 40% opaque, therefore 60% transparency.}

\item{label_style}{`character` string indicating the style of label
to display:
`"basic"` displays text with no background shading or border,
`"fill"` displays text on opaque colored background,
`"shaded"` displays text on partially transparent colored background,
`"lite"` displays text on partially transparent lite background,
`"lite_box"` displays text on lite background with border.}

\item{...}{additional arguments are passed to `list2imsigned()`.}
}
\description{
Directional Venn diagram
}
\details{
This function displays a Venn diagram, or when `proportional=TRUE`
it displays a Euler diagram, representing counts in each overlap
set.

Input data is supplied either as a `list` or `incidence matrix`,
or `signed incidence matrix` (whose values indicate direction),
and is passed to `signed_overlaps()` to summarize counts by
Venn overlaps.

By default, when input data contains signed direction, the
counts include summary counts for the different forms of
agreement in direction. See `signed_overlaps()` for description
of each `overlap_type`, for different methods of summarizing
the overlapping directions.
}
\examples{
setlist <- make_venn_test(100, 3);
venndir(setlist)

setlist <- make_venn_test(100, 3, do_signed=TRUE);
venndir(setlist)
venndir(setlist)

venndir(setlist, label_style="basic")
venndir(setlist, label_style="fill")
venndir(setlist, label_style="shaded")
venndir(setlist, label_style="shaded_box")
venndir(setlist, label_style="lite")
venndir(setlist, label_style="lite_box")

# proportional Euler diagram
venndir(setlist,
   proportional=TRUE);

# nudge circles and hide the zero
venndir(setlist,
   proportional=TRUE,
   display_zero=FALSE,
   x_nudge=c(set_C=-0.4),
   y_nudge=c(set_C=0.1))

}
\seealso{
Other venndir overlaps: 
\code{\link{signed_overlaps}()},
\code{\link{textvenn}()}
}
\concept{venndir overlaps}
