% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-poly-conversions.R
\name{xy_list_to_coord_list}
\alias{xy_list_to_coord_list}
\title{Convert xy_list to coordinate list format}
\usage{
xy_list_to_coord_list(x, closed = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{\code{list} of coordinates with elements \code{"x"} and \code{"y"} which
may themselves each be a \code{list} representing multiple polygons.}

\item{closed}{\code{logical} indicating whether the last point in each
polygon should equal the first point in the polygon, to "close"
the path so to speak.}

\item{...}{additional arguments are ignored.}
}
\description{
Convert xy_list to coordinate list format
}
\examples{
counts <- c(A=7, B=0, `A&B`=1)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
plot_polygon_list(polygon_list, col=2:4)

A_only <- minus_polygon_list(polygon_list, new_name="A_only");
xy_list <- polygon_list_to_xy_list(A_only)
coord_list <- xy_list_to_coord_list(xy_list)
jamba::nullPlot(doBoxes=FALSE, xlim=c(-4, 4), ylim=c(-4, 4), asp=1)
polypath(x=coord_list$x, y=coord_list$y, rule="evenodd", col=c("red", "blue", "gold"))

counts <- c(A=7, B=0, C=0, `A&B`=1, `A&C`=2)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
plot_polygon_list(polygon_list, col=2:4)

A_only <- minus_polygon_list(polygon_list, new_name="A_only");
xy_list <- polygon_list_to_xy_list(A_only)
coord_list <- xy_list_to_coord_list(xy_list, closed=TRUE)
jamba::nullPlot(doBoxes=FALSE, xlim=c(-4, 4), ylim=c(-4, 4), asp=1)
polypath(x=coord_list$x, y=coord_list$y, rule="evenodd", col=c("red", "blue", "gold"))

plot_polygon_list(c(polygon_list[1]),
   col=jamba::alpha2col(alpha=0.5, c("red", "blue", "gold")))
plot_polygon_list(c(A_only, polygon_list[2:3])[1],
   col=jamba::alpha2col(alpha=0.5, c("red", "blue", "gold")))

# show proper label avoiding the hole
points(polygon_list_labelr(c(A_only)), pch=20, col="red")
}
\seealso{
Other venndir polygons: 
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_polygon_buffer}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{union_polygon_list}()}
}
\concept{venndir polygons}
