% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-poly-conversions.R
\name{xy_list_to_polygon_list}
\alias{xy_list_to_polygon_list}
\title{Convert coordinate list of x and y into polygon list of x,y coordinates}
\usage{
xy_list_to_polygon_list(x, ...)
}
\description{
Convert coordinate list of x and y into polygon list of x,y coordinates
}
\examples{
counts <- c(A=1, B=2, `A&B`=3)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
xy_list <- polygon_list_to_xy_list(polygon_list)
polygon_list2 <- xy_list_to_polygon_list(xy_list)
identical(polygon_list, polygon_list2)

}
