% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{add_orientation_JamPolygon}
\alias{add_orientation_JamPolygon}
\title{Define orientation for each polygon, clockwise or counterclockwise}
\usage{
add_orientation_JamPolygon(
  jp,
  flip_sign = 1,
  include_holes = FALSE,
  include_clockwise = FALSE,
  include_parent = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{jp}{\code{JamPolygon}}

\item{flip_sign}{\code{integer} to indicate whether the orientation should
be flipped, inverting the meaning of all returned values.
For example, clockwise polygons by default are considered
"solid", and counter-clockwise are considered "holes", for
the purpose of interpreting nested polygons. When \code{flip_sign=-1}
it reverses these assumptions.
This option is experimental and intended mainly for internal
testing.}

\item{include_holes, include_clockwise, include_parent}{\code{logical}
indicating whether to include additional columns to describe
multi-part polygons. The clockwise direction is used to define
the inner and outer edge of the border, usually defined as
left or right, in relation to the direction the line is drawn.
Nested multi-part polygons define interior holes, and are distinguished
from adjacent polygons.
\itemize{
\item \code{"polygon_clockwise"}: 1 clockwise, -1 counter-clockwise
\item \code{"polygon_holes"}: 1 is a solid polygon, -1 is a hole inside
another polygon indicated with \code{"polygon_parent"}.
\item \code{"polygon_parent"}: integer index used to indicate the parent
polygon for nested multi-part polygons. The index will indicate
itself, when the polygon is not nested inside another polygon.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored}
}
\value{
\code{JamPolygon} with column \code{"orientation"} added to
slot "polygons".
}
\description{
Define orientation for each polygon, clockwise or counterclockwise
}
\details{
Todo:
\itemize{
\item When polygons are nested, which is intended to create holes inside
one larger polygon, the hole should have its orientation flipped
so that the border is drawn on the opposite side per its orientation.
So the task is to determine whether a polygon is the hole of
another polygon.
}
}
\examples{
df3 <- data.frame(name=c("polygon1", "polygon2"),
   label=c("polygon1", "polygon2"),
   x=I(list(
      list(c(1, 6, 6, 1),
         c(2, 5, 5, 2),
         c(3, 4, 4, 3)),
      list(#c(11, 16, 16, 11),
         c(12, 15, 15, 12),
         c(13, 14, 14, 13))
      )),
   y=I(list(
      list(c(1, 1, 6, 6),
         c(2, 2, 5, 5),
         c(3, 3, 4, 4)),
      list(#c(1, 1, 6, 6),
         c(2, 2, 5, 5),
         c(3, 3, 4, 4))
      )),
   fill=c("gold", "firebrick"))
jp3 <- new("JamPolygon", polygons=df3);
plot(jp3);

add_orientation_JamPolygon(jp3,
   include_holes=TRUE,
   include_clockwise=TRUE,
   include_parent=TRUE)

}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon,ANY,ANY,ANY-method}},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
