% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{buffer_JamPolygon}
\alias{buffer_JamPolygon}
\title{Apply buffer outside or inside JamPolygon}
\usage{
buffer_JamPolygon(
  jp,
  buffer = -0.5,
  steps = 50,
  relative = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{jp}{\code{JamPolygon} with one or more polygons. When multiple polygons
are provided, they are combined with \code{union_JamPolygon()} so that
one overall buffer can be provided.}

\item{buffer}{\code{numeric} buffer, where negative values cause the polygon
to be reduced in size.}
}
\description{
Apply buffer outside or inside JamPolygon
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{plot,JamPolygon,missing-method}},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
