% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{check_JamPolygon}
\alias{check_JamPolygon}
\title{Check JamPolygon object}
\usage{
check_JamPolygon(object)
}
\arguments{
\item{object}{\code{JamPolygon} object}
}
\description{
Check whether a JamPolygon object is valid.
}
\details{
It requires:
\itemize{
\item \code{slotName(object)} to contain \code{"polygons"}
\item \code{object@polygons} to inherit \code{"data.frame"}
\item \code{colnames(object@polygons)} to contain \code{c("name", "x", "y")}
}

General guidance for JamPolygon objects:
\itemize{
\item Empty polygons can be represented as one row in the \code{data.frame}
which contains \code{NULL} or \code{numeric(0)} x,y coordinates.
\itemize{
\item Empty polygons can be the result of \code{intersect_JamPolygon()}
that finds no suitable intersect.
\item Note that downstream processing, and plotting, must account for
empty polygons and handle or ignore them accordingly.
}
}
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
