% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndir-class.R
\name{check_Venndir}
\alias{check_Venndir}
\title{Check Venndir object}
\usage{
check_Venndir(object)
}
\arguments{
\item{object}{\code{Venndir} object}
}
\description{
Check Venndir object integrity
}
\details{
\subsection{Todo:}{
\itemize{
\item Write validation function.
\item Consider whether to have \code{jp} (sets), \code{jps} (overlaps) in separate ojbects
}
}
}
