% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-render.R
\name{degrees_to_adj}
\alias{degrees_to_adj}
\title{Degrees to text adjustment}
\usage{
degrees_to_adj(degrees, top = 90, clockwise = FALSE, expand = 0, ...)
}
\arguments{
\item{degrees}{\code{numeric} value for angles in degrees}

\item{top}{\code{numeric} value indicating the angle at the
top position}

\item{clockwise}{\code{logical} indicating whether the angle
increases in clockwise direction}

\item{expand}{\code{numeric} value intended to expand the adjust
value. For example \code{expand=0.5} will expand the adjust
value 50\%.}

\item{...}{additional arguments are ignored}
}
\description{
Degrees to text adjustment
}
\details{
Utility function to define \code{adj} values suitable
for text plotting, which arranges text relative
to the angle in degrees.
}
\examples{
degrees <- seq(from=1, to=360, by=33);
adjdf <- degrees_to_adj(degrees);
x <- cos(jamba::deg2rad(degrees));
y <- sin(jamba::deg2rad(degrees));
plot(x, y,
   pch=20, asp=1,
   xlim=c(-1.3, 1.3),
   ylim=c(-1.3, 1.3));
for (i in seq_along(degrees)) {
   text(labels=i,
      x=x[i], y=y[i],
      adj=unlist(adjdf[i,]))
}

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{diff_degrees}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
