% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-angles.R
\name{display_angles}
\alias{display_angles}
\title{Display degree angles around a unit circle}
\usage{
display_angles(x, add = FALSE, col = "darkorange", lwd = 2, ...)
}
\arguments{
\item{x}{\code{numeric} angles in degrees}

\item{add}{\code{logical} whether to add to an existing plot, default FALSE.}

\item{col}{\code{character} R color to use for arrows, default 'darkorange'.}

\item{lwd}{\code{numeric} line width, default 2.}

\item{...}{additional arguments are ignored.}
}
\description{
Display degree angles around a unit circle
}
\examples{
display_angles(jamba::nameVector(c(10, 45, 90, 225)))

}
\seealso{
Other venndir geometry: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{rescale_coordinates}()},
\code{\link{spread_degrees}()},
\code{\link{three_point_angle}()}
}
\concept{venndir geometry}
