% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-labels.R
\name{draw_gridtext_groups}
\alias{draw_gridtext_groups}
\title{Draw boundary around groups of gridtext labels}
\usage{
draw_gridtext_groups(
  g_labels,
  gdf,
  groupdf = NULL,
  segment_df = NULL,
  realign = TRUE,
  adjust_center = FALSE,
  adjx_fn = c,
  adjy_fn = c,
  do_draw = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g_labels}{\code{richtext_grob} class as returned by
\code{gridtext::richtext_grob()}.}

\item{gdf}{\code{data.frame} with one row per \code{grob} in \code{g_labels},
with colnames \code{"x"}, \code{"y"}, \code{"overlap_set"}, \code{"r"},
\code{"r_unit"}, \code{"border_col"}, \code{"box_fill"}. The rows
are grouped using \code{c("x", "y", "overlap_set")}.
The \code{"r"} and \code{"r_unit"} values are used to define
the radius and radius unit for the rounded corner.
The row in \code{gdf} with the highest \code{"r"} value is the
reference row, for the \code{"r"} radius, and the \code{"border_col"}
and \code{"box_fill"} values.}

\item{groupdf}{\code{data.frame} with information about label groupings,
with columns: overlap_set, counttype, type, location, roworder,
label_df_rowname, name, childName, gdf_group.
Importantly, it must be in identical row order as \code{gdf}.}

\item{segment_df}{\code{data.frame} with segment coordinates, optional.}

\item{realign}{\code{logical} experimental feature indicating whether to
realign grouped labels, thus ignoring the hjust/vjust for each
label, instead arranging the labels using \code{groupdf}.
In principle, overlap label (set name) should be top-center,
and counts should be centered below. When there are main counts
and signed counts, they should be in two columns, collectively
centered below the main label.}

\item{adjust_center}{\code{logical} indicating whether to adjust the x-centering
for a group of labels when \code{adjx=0.5}, using the total width of the
group of labels. When FALSE (default) the label x-position is not
adjusted, which assumes labels are already oriented properly with
respect to the x-axis position.
This option is \code{FALSE} specifically for two-column labels positioned
around a fixed centerpoint, where one column could be wider than the
other, and it would otherwise push the other column the other way.}

\item{do_draw}{\code{logical} indicating whether to draw the finished
result in the context of the current open graphics device.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Draw boundary around groups of gridtext labels
}
\details{
This function is an internal function used to take individual
labels defined by \code{gridtext::richtext_grob()}, and creating
\code{grid::roundrectGrob()} objects that cover each group of
labels with one rounded rectangular boundary.
}
\seealso{
Other venndir label: 
\code{\link{reposition_venn_gridtext_labels}()}
}
\concept{venndir label}
