% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{eulerr2polys}
\alias{eulerr2polys}
\title{Convert euler output to polygons}
\usage{
eulerr2polys(x)
}
\arguments{
\item{x}{output from \code{eulerr::euler()}}
}
\value{
\code{sp::SpatialPolygons} object with one \code{sp::Polygons}
element for each Euler circle or ellipse.
}
\description{
Convert euler output to polygons
}
\details{
This function takes the output from \code{eulerr::euler()} and
converts it to polygons in \code{sp::SpatialPolygons} format.
Essentially the output from \code{eulerr::euler()} is either a
set of circles, or ellipses.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_venndir_label}()},
\code{\link{print_color_df}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{shrink_df}()},
\code{\link{sp_percent_area}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
