% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{eulerr2polys}
\alias{eulerr2polys}
\title{Convert eulerr() output to polygons}
\usage{
eulerr2polys(x)
}
\description{
Convert eulerr() output to polygons
}
\examples{
if (require(eulerr)) {
}

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_sp}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{union_polygons}()}
}
\concept{venndir utility}
\section{@import sp}{
}
