% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{eulerr2polys}
\alias{eulerr2polys}
\title{Convert euler output to polygons}
\usage{
eulerr2polys(x)
}
\value{
\code{sp::SpatialPolygons} object with one \code{sp::Polygons}
element for each Euler circle or ellipse.
}
\description{
Convert euler output to polygons
}
\details{
This function takes the output from \code{eulerr::euler()} and
converts it to polygons in \code{sp::SpatialPolygons} format.
Essentially the output from \code{eulerr::euler()} is either a
set of circles, or ellipses.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
