% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{eulerr_to_polygon_list}
\alias{eulerr_to_polygon_list}
\title{Convert euler output to polygons}
\usage{
eulerr_to_polygon_list(x)
}
\arguments{
\item{x}{output from \code{eulerr::euler()}}
}
\value{
\code{list} polygon object with one polygon
for each Euler circle or ellipse.

\code{list} with polygons for each unique set defined by \code{names(x)},
where each list contains \code{numeric} vectors named \code{"x"} and \code{"y"}.
}
\description{
Convert euler output to polygons
}
\details{
This function takes the output from \code{eulerr::euler()} and
converts it to polygons in \code{list} format.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)

polygon_areas(polygon_list)

}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
