% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-render.R
\name{expand_range}
\alias{expand_range}
\title{Expand numeric range}
\usage{
expand_range(x, expand_fraction = 0.1, minimum_range = 0.01, ...)
}
\arguments{
\item{x}{\code{numeric} vector, or \code{list} of \code{numeric} vectors.
The input is converted to a range using \code{range(x, na.rm=TRUE)}
to ensure no \code{NA} values are included. Note that this
step will force the range to be in ascending order.}

\item{expand_fraction}{\code{numeric} value indicating the
fraction of the range defined by \code{diff(range(x, na.rm=TRUE))}
to add to the total range. When \code{expand_fraction} has
only one value, it is applied across the whole range,
which means each side is extended by half the \code{expand_fraction}.
When \code{expand_fraction} contains two values, it is
applied in order to the low, then high side of the
numeric range, and each full \code{expand_range} value is
applied.}

\item{minimum_range}{\code{numeric} value indicating the minimum
range of the output, useful when the input has zero
range, for example if \code{x=c(10, 10)}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} vector, or when input \code{x} is \code{list} the
output will also be a \code{list} of \code{numeric} vectors.
The numeric vector will contain the range after
expansion.
}
\description{
Expand numeric range
}
\details{
This function takes a \code{numeric} range (or numeric vector
and calculates its range) and expands this range by
a fraction given by \code{expand_fraction}.

When the input range is zero, the minimum absolute range
can be given by \code{minimum_range}.

The input may be a \code{list} that contains \code{numeric} vectors,
in which case the \code{list} will be iterated to produce an
expanded range for each \code{numeric} vector. Each \code{numeric}
vector is expanded independently.

This function is intended to be a simple technique to expand
x-axis and y-axis ranges of a graphical plot.
}
\examples{
x <- c(0, 10);

# expand the total range by 0.1
expand_range(x, 0.1);

# the original range is 10 units
diff(x);

# the expanded range is 11 units
diff(expand_range(x, 0.1));

# expand one side but not the other
expand_range(x, c(0.1, 0));
# this new range is 11 units
diff(expand_range(x, c(0.1, 0)))

# input with no range is extended to some minimum value
expand_range(1, minimum_range=1)
expand_range(1, minimum_range=c(1, 0))

# list input is iterated, for example xlim, ylim
xlim <- c(1, 10)
ylim <- c(1, 100)
expand_range(list(xlim=xlim, ylim=ylim))

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{eulerr2polys}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_percent_area}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
