% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{find_vennpoly_overlaps}
\alias{find_vennpoly_overlaps}
\title{Find Venn polygon overlaps}
\usage{
find_vennpoly_overlaps(
  sp,
  venn_counts = NULL,
  venn_items = NULL,
  venn_colors = NULL,
  preset = "dichromat",
  x_nudge = NULL,
  y_nudge = NULL,
  do_plot = FALSE,
  verbose = FALSE,
  ...
)
}
\value{
\code{sp::SpatialPolygonsDataFrame} object, with
additional columns:
\itemize{
\item \code{"label"}
\item \code{"venn_counts"}
\item \code{"venn_color"}
\item \code{"x_label"}
\item \code{"y_label"}
}
}
\description{
Find Venn polygon overlaps
}
\details{
This function takes a named list of polygons
and returns the combination of polygon overlaps
as used in a Venn diagram.

When a vector of Venn counts is supplied, the
counts are associated with the respective polygon,
and any counts not represented by a polygon
are returned as an attribute \code{"venn_missing"}.
}
