% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{get_largest_polygon}
\alias{get_largest_polygon}
\title{Largest polygon in a SpatialPolygons object}
\usage{
get_largest_polygon(sp, ...)
}
\arguments{
\item{sp}{object with class \code{sp::SpatialPolygons} that may
contain one or more closed polygons.}

\item{...}{additional arguments are ignored.}
}
\description{
Largest polygon in a SpatialPolygons object
}
\details{
This function returns the largest polygon in a
\code{sp::SpatialPolygons} object, notably when there
are multiple polygons contained in one object.
The function calculates the largest area using
\code{rgeos:gArea()}.

If two polygons have identical area, the first
polygon is returned.

This function calls a helper function \code{sp::disaggregate()}
which creates one \code{sp::SpatialPolygons} object
for each individual polygon in the source object.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_sp_buffer}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_percent_area}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
