% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{get_largest_polygon}
\alias{get_largest_polygon}
\title{Largest polygon in a SpatialPolygons object}
\usage{
get_largest_polygon(sp, ...)
}
\description{
Largest polygon in a SpatialPolygons object
}
\details{
This function returns the largest polygon in a
`sp::SpatialPolygons` object, notably when there
are multiple polygons contained in one object.
The function calculates the largest area using
`rgeos:gArea()`.

If two polygons have identical area, the first
polygon is returned.
}
