% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-shapes.R
\name{get_venn_polygon_shapes}
\alias{get_venn_polygon_shapes}
\title{Get Venn shapes as polygon_list}
\usage{
get_venn_polygon_shapes(
  counts,
  proportional = FALSE,
  sep = "&",
  circles_only = FALSE,
  circle_nudge = NULL,
  rotate_degrees = 0,
  return_type = c("JamPolygon"),
  seed = 123,
  ...
)
}
\arguments{
\item{counts}{\code{integer} vector whose names represent set overlaps,
where set names are separated by \code{sep} delimiter.}

\item{proportional}{\code{logical} indicating whether to create proportional
circles, where \code{proportional=FALSE} creates standard Venn diagram,
and \code{proportional=TRUE} creates a Euler diagram.}

\item{sep}{\code{character} delimiter used to separate set names in
\code{names(counts)}.}

\item{circles_only}{\code{logical} indicating whether to force Venn
4-way diagram to use only circles; or passed to \code{eulerr::euler()}
to force it to return circles instead of allowing ellipse shapes.}

\item{circle_nudge}{\code{list} of \code{numeric} vectors each length 2, whose
names match set names derived from \code{counts}. For example if
\code{counts=c(set_A=5, set_B=10, "setA&set_B"=3)}, then to nudge
the \code{set_A} circle, use \code{circle_nudge=list(set_A=c(1, 0))}.
This argument is intended to allow manipulation of specific
circle or ellipse positions for aesthetic effects. Particularly
for proportional Euler diagrams, sometimes the algorithm places
circles in non-ideal locations}

\item{rotate_degrees}{\code{numeric} value indicating rotation in degrees
for the entire set of shapes. This argument is intended to
change the overall orientation, for example so that certain
sets are at the top.}

\item{return_type}{\code{character} with only one permitted value 'JamPolygon',
and only present for brief backward compatibility.}

\item{seed}{\code{numeric} default 123, passed to \code{set.seed()} when calling
\code{eulerr::euler()}. When \code{NULL} it is not set, which permits fully
random output.}

\item{...}{additional arguments are passed to internal functions,
notably \code{eulerr::euler()} which offers detailed options for
optimization of the returned Euler shapes. Note that for two sets,
optimization is mostly irrelevant, but three shapes, and/or ellipse
shapes, optimization could be adjusted.
Also see \code{eulerr::eulerr_options()} for an alternative method
to customize the Euler optimization.}
}
\value{
\code{JamPolygon} representing Venn or Euler circles,
however this function does not subdivide circles by overlap.
}
\description{
Get Venn shapes as polygon_list
}
\details{
This function takes a Venn overlap counts and creates
corresponding circles or ellipses that represent
either a Venn diagram, or proportional Venn (Euler)
diagram.

For non-proportional Venn diagrams, this function accepts
up to 5 sets, although the 5-way Venn diagram is not
visually intuitive.

For proportional Euler diagrams, this function simply passes
the count vector to \code{eulerr::euler()} and returns the output.
That function accepts more sets, however not all overlaps may
be represented in the output.
}
\examples{
counts <- c(A=1, B=2, `A&B`=3, C=4)
venn_colors <- colorjam::rainbowJam(3, alpha=0.5);

vjp <- get_venn_polygon_shapes(counts, return_type="JamPolygon")
plot(vjp, fill=venn_colors)

vjp <- get_venn_polygon_shapes(counts,
   return_type="JamPolygon",
   proportional=TRUE)
plot(vjp, fill=venn_colors)

}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{check_systemfonts_family}()},
\code{\link{expand_range}()},
\code{\link{hexsticker_venndir}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{subset_systemfonts}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
