% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-shapes.R
\name{get_venn_shapes}
\alias{get_venn_shapes}
\title{Get Venn shapes}
\usage{
get_venn_shapes(
  counts,
  proportional = FALSE,
  sep = "&",
  circles_only = FALSE,
  circle_nudge = NULL,
  rotate_degrees = rotate_degrees,
  ...
)
}
\description{
Get Venn shapes
}
\details{
This function takes a Venn overlap counts
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{make_venn_test}()},
\code{\link{nudge_sp}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()}
}
\concept{venndir utility}
