% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{intersect_JamPolygon}
\alias{intersect_JamPolygon}
\title{Intersect one or more JamPolygon objects}
\usage{
intersect_JamPolygon(jp, new_name = NULL, ...)
}
\arguments{
\item{jp}{\code{JamPolygon}}

\item{new_name}{\code{character} string used to populate the \code{"name"}
column in the output data. Only the first value is used,
otherwise the first existing value in \code{names(jp)} is kept.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{JamPolygon} after applying the intersection
}
\description{
Intersect one or more JamPolygon objects
}
\examples{
df3 <- data.frame(name=c("polygon1", "polygon2"),
   label=c("polygon1", "polygon2"),
   x=I(list(
      list(c(1, 6, 6, 1),
         c(2, 5, 5, 2),
         c(3, 4, 4, 3)),
      list(
         c(12, 15, 15, 12) - 7.5,
         c(13, 14, 14, 13) - 7.5)
   )),
   y=I(list(
      list(c(1, 1, 6, 6),
         c(2, 2, 5, 5),
         c(3, 3, 4, 4)),
      list(
         c(2, 2, 5, 5),
         c(3, 3, 4, 4))
   )),
   fill=jamba::alpha2col(c("gold", "firebrick"), alpha=0.7))
jp3 <- new("JamPolygon", polygons=df3);
plot(jp3);

jp3i <- intersect_JamPolygon(jp3)
jp3i@polygons$fill <- "red"
jp3i@polygons$border <- "red3";
jp3i@polygons$border.lwd <- 3;
jp3c <- rbind2(jp3, jp3i)
plot(jp3c)
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon,ANY,ANY,ANY-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
