% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{intersect_polygons}
\alias{intersect_polygons}
\title{Intersect one or more polygons}
\usage{
intersect_polygons(sp, ...)
}
\arguments{
\item{sp}{\code{list} object that contains one or more
\code{sp::SpatialPolygons}.}

\item{...}{additional arguments are ignored.}
}
\value{
object \code{sp::SpatialPolygons}
}
\description{
Intersect one or more polygons
}
\details{
This function takes a \code{list} of \code{sp::SpatialPolygons}
objects, and performs iterative \code{rgeos::gIntersection()} on
the full list, since \code{rgeos::gIntersection()} only operates on
two polygons at a time.
}
\examples{
circles <- get_venn_shapes(c(A=1, B=1, C=1))
col3 <- c("#FF000077", "#DDDD0088", "#0000FF77")
sp::plot(circles, col=col3)

circle_intersect <- intersect_polygons(circles);
plot(circle_intersect, add=TRUE, col="gold", lwd=3);

}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_sp}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()}
}
\concept{venndir spatial}
