% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-label-outside-jp.R
\name{label_outside_JamPolygon}
\alias{label_outside_JamPolygon}
\title{Position labels outside JamPolygon}
\usage{
label_outside_JamPolygon(
  jp,
  which_jp = NULL,
  center = NULL,
  buffer = 0.1,
  center_method = c("bbox", "label"),
  vector_method = c("label", "farthest"),
  segment_method = c("nearest", "vector"),
  distance = 0.1,
  min_degrees = 15,
  relative = TRUE,
  seed = 123,
  debug = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{jp}{\code{JamPolygon}}

\item{which_jp}{\code{integer} or \code{NULL}; when \code{which_jp} contains one
or more \code{integer} values, they refer to rows in \code{jp}, and
each will be analyzed in sequence. When \code{which_jp=NULL} then
all the polygons in \code{jp} will be analyzed in sequence.}

\item{center}{\code{numeric} vector or matrix with two values indicating
the center position. When \code{center=NULL} then the center is
determined using a method defined by \code{center_method}.}

\item{center_method}{\code{character} string indicating the method to
determine the \code{center}:
\itemize{
\item \code{"label"} uses the mean x,y coordinate
of all the polygon label positions;
\item \code{"bbox"} uses the mean x,y
coordinate of the bounding box that encompasses the polygons.
}

The effect
is to extend outer labels radially around this center point.
Using the mean label position with \code{center_method="label"}
is helpful because it ensures labels are extended in all directions
even when most labels are in one upper or lower section of
the \code{sp} polygons.}

\item{distance}{\code{numeric} value indicating the absolute coordinate
distance away from the perimiter of \code{jp} to place labels.
This value is the \code{buffer} for \code{buffer_JamPolygon()}.}
}
\description{
Position labels outside JamPolygon
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{plot,JamPolygon,missing-method}},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
