% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-label-outside-jp.R
\name{label_segment_JamPolygon}
\alias{label_segment_JamPolygon}
\title{Define a label segment for JamPolygon}
\usage{
label_segment_JamPolygon(
  x0,
  x1,
  y0,
  y1,
  jp,
  return_class = c("point", "matrix"),
  buffer = 0,
  plot_debug = FALSE,
  relative = TRUE,
  verbose = FALSE,
  ...
)
}
\description{
Define a label segment for JamPolygon
}
