% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-label-outside-jp.R
\name{label_segment_JamPolygon}
\alias{label_segment_JamPolygon}
\title{Define a label segment for JamPolygon}
\usage{
label_segment_JamPolygon(
  x0,
  x1,
  y0,
  y1,
  jp,
  return_class = c("point", "matrix"),
  buffer = 0,
  plot_debug = FALSE,
  relative = TRUE,
  verbose = FALSE,
  ...
)
}
\description{
Define a label segment for JamPolygon
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{plot,JamPolygon,missing-method}},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
