% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{labelr_JamPolygon}
\alias{labelr_JamPolygon}
\title{Define label positions for JamPolygon using polylabelr::poi()}
\usage{
labelr_JamPolygon(
  jp,
  precision = 1,
  add_to_jp = FALSE,
  subset_multipart = TRUE,
  ...
)
}
\arguments{
\item{jp}{\code{JamPolygon}}

\item{precision}{\code{numeric} passed to \code{polylabelr::poi()}, default 1.}

\item{add_to_jp}{\code{logical} default FALSE, indicating whether to
add \code{"label_x"},\code{"label_y"} into \code{jp@polygons} for persistence.}

\item{subset_multipart}{\code{logical} default TRUE, whether to subset
multipart polygons to label the largest area.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{matrix} with columns \code{"x"}, \code{"y"}, \code{"dist"}
}
\description{
Define label positions for JamPolygon using polylabelr::poi()
}
\details{
This function is a simple wrapper for \code{polylabelr::poi()}, which
recognizes holes inside polygons.
\subsection{Todo:}{
\itemize{
\item Consider basic support for non-overlapping label positions.
\enumerate{
\item First pass might be to use non-overlapping regions of a polygon
among a set of polygons.
\item Second pass would be to use the region with fewest overlaps
among other polygons in the set, to define the label position.
This rule could serve to solve (1) as well.
}
}
}
}
\examples{
df3 <- data.frame(name=c("polygon1", "polygon2"),
label=c("polygon A", "polygon B"),
x=I(list(
   list(c(1, 6, 6, 1),
      c(2, 5, 5, 2),
      c(3, 4, 4, 3)),
   list(
      c(12, 15, 15, 12) - 3,
      c(13, 14, 14, 13) - 3)
)),
y=I(list(
   list(c(1, 1, 6, 6),
      c(2, 2, 5, 5),
      c(3, 3, 4, 4)),
   list(
      c(2, 2, 5, 5),
      c(3, 3, 4, 4))
)),
fill=jamba::alpha2col(c("gold", "firebrick"), alpha=0.7))
jp3 <- new("JamPolygon", polygons=df3);
plot(jp3);

}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon,ANY,ANY,ANY-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
