% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-im.R
\name{list2im_opt}
\alias{list2im_opt}
\title{Optimized list to incidence matrix}
\usage{
list2im_opt(setlist, empty = 0, do_sparse = FALSE, ...)
}
\arguments{
\item{setlist}{\code{list} of vectors}

\item{empty}{default single value used for empty/missing entries,
the default \code{empty=0} uses zero for entries not present.
Another alternative is \code{NA}.
Providing a \code{character} value will convert the output to
a \code{character} matrix, be warned.}

\item{do_sparse}{\code{logical} indicating whether to coerce the output
to sparse matrix class \code{"CsparseMatrix"} from the Matrix package.
The default is \code{FALSE} as of version 0.0.33.900, since the
most common use case requires a regular matrix.
For extremely large data, consider using a sparse matrix.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{matrix} object with value \code{c(0, 1)} when \code{do_sparse=FALSE}
(default), or when \code{do_sparse=TRUE}, it returns a \code{Matrix} object
class \code{"CsparseMatrix"} with \code{logical} values, only when
\code{Matrix} is available.
}
\description{
Optimized conversion of list to incidence matrix
}
\details{
This function rapidly converts a list of vectors into
an incidence matrix whose rownames are items, and colnames
are the names of the input list. The default output
\code{do_sparse=TRUE} returns a \code{logical} matrix class \code{ngCMatrix}
from the \code{Matrix} package. When \code{do_sparse=FALSE} the
output is a \code{matrix} class with \code{numeric} values \code{0} and \code{1}.

Note that the rows in the output matrix are not sorted,
since the step of sorting item names may take several
seconds when working with a list whose vectors contain
millions of items. For sorted rows, the best remedy is
to run this function, the re-order rownames afterward.
}
\examples{
setlist <- list(A=c("one", "two", "three"),
   b=c("two", "one", "four", "five"));
list2im_opt(setlist);

list2im_opt(setlist, do_sparse=TRUE);

}
\seealso{
Other venndir conversion: 
\code{\link{collapse_im}()},
\code{\link{counts2setlist}()},
\code{\link{im2list}()},
\code{\link{im_value2list}()},
\code{\link{list2im_value}()},
\code{\link{overlaplist2setlist}()},
\code{\link{signed_counts2setlist}()}
}
\concept{venndir conversion}
