% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-im.R
\name{list2im_opt}
\alias{list2im_opt}
\title{Optimized list to incidence matrix}
\usage{
list2im_opt(setlist, empty = 0, do_sparse = TRUE, ...)
}
\arguments{
\item{setlist}{\code{list} of vectors}

\item{empty}{default single value used for empty/missing entries,
the default \code{empty=0} uses zero for entries not present.
Another alternative is \code{NA}.}

\item{do_sparse}{\code{logical} indicating whether to coerce the output
to sparse matrix class \code{"ngCMatrix"} from the Matrix package.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{Matrix} object with class \code{"ngCMatrix"} that contains
contains \code{logical} values. When \code{do_sparse=FALSE} the returned
object is \code{matrix} with values \code{c(0, 1)}.
}
\description{
Optimized list to incidence matrix
}
\details{
This function rapidly converts a list of vectors into
an incidence matrix with value of \code{1} for each entry
(row) present in each input list (column).

Note that the rows in the output matrix are not sorted,
since this step can take several seconds when working with
a list whose vectors contain millions of rows.
}
\seealso{
Other venndir conversion: 
\code{\link{counts2setlist}()},
\code{\link{im2list}()},
\code{\link{im_value2list}()},
\code{\link{list2im_value}()},
\code{\link{overlaplist2setlist}()},
\code{\link{signed_counts2setlist}()}
}
\concept{venndir conversion}
