% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-printcolordf.R
\name{make_color_contrast}
\alias{make_color_contrast}
\title{Make color contrast with background colors}
\usage{
make_color_contrast(
  x,
  y = NULL,
  bg = "white",
  L_threshold = 65,
  C_floor = 90,
  L_hi = 95,
  L_lo = 40,
  blend_preset = "ryb",
  do_plot = FALSE,
  cex = 2,
  ...
)
}
\arguments{
\item{x}{\code{character} vector of text colors}

\item{y}{\code{character} vector of fill colors for which the colors
in \code{x} should be adjusted for visual contrast.}

\item{bg}{\code{character} background color, default 'white', used
when colors in \code{y} contain transparency.}

\item{L_threshold}{\code{numeric} threshold Luminance (L) value which
defines the point where the background color should have
light or dark text for optimum visual contrast. The default \code{65}
is well-discussed in data visualization forums, however
may not be ideal for all viewing conditions.}

\item{C_floor}{\code{numeric} to define the minimum Chroma (C) when
the adjusted color Chroma is below an internal threshold 20.}

\item{L_hi, L_lo}{\code{numeric} value to define the Luminance (L) used
for 'hi' (bright) output colors, or 'lo' (dark) output
colors. The default 'L_hi=95' is fairly high, and 'L_lo=40'
is moderately dark. In future the recommended default may
become 'L_hi=85' to improve the resulting color hue.}

\item{blend_preset}{\code{character} string passed as \code{preset} to
\code{colorjam::blend_colors()} to define the color wheel used
during color blending operations. Default 'ryb' uses the
red-yellow-blue color wheel. This adjustment only affects
layering 'y' together with 'bg'.}

\item{do_plot}{\code{logical} whether to create a visual plot illustrating
the input and output colors, on top of the background.}

\item{cex}{\code{numeric} text size adjustment used when \code{do_plot=TRUE}.}

\item{...}{additional arguments are passed to internal functions,
including \code{colorjam::blend_colors()}.}
}
\description{
Make color contrast with background colors
}
\details{
This function provides a simple method to ensure a color
has adequate visual contrast with a background color,
while retaining some color saturation. For example,
red on red background will return something close to
pink, so the pink retains the red color saturation
but is visually distinct from the background color red.

Similarly, pink on pink will return something close to
red.
}
\examples{
x <- c("firebrick", "dodgerblue", "gold", "pink", "white");
y <- rep("red3", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- rep("pink1", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- rep("gold", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- c("red4", "aquamarine4", "blue3", "yellow", "pink2")
make_color_contrast(x, y, do_plot=TRUE, C_floor=140);

}
\seealso{
Other venndir advanced: 
\code{\link{collapse_im}()},
\code{\link{combine_marquee_styles}()},
\code{\link{nudge_venndir_label}()}
}
\concept{venndir advanced}
