% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-printcolordf.R
\name{make_color_contrast}
\alias{make_color_contrast}
\title{Make color contrast}
\usage{
make_color_contrast(
  x,
  y = NULL,
  bg = "white",
  L_threshold = 65,
  C_floor = 130,
  L_hi = 100,
  L_lo = 40,
  do_plot = FALSE,
  cex = 2,
  ...
)
}
\description{
Make color contrast
}
\details{
This function provides a simple method to ensure a color
has adequate visual contrast with a background color,
while retaining some color saturation. For example,
red on red background will return something close to
pink, so the pink retains the red color saturation
but is visually distinct from the background color red.

Similarly, pink on pink will return something close to
red.
}
\examples{
x <- c("red", "blue", "gold", "ivory", "pink", "white");
y <- rep("red4", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- rep("pink1", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- rep("gold", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- c("red4", "aquamarine4", "blue3", "yellow", "pink2")
make_color_contrast(x, y, do_plot=TRUE, C_floor=140);

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_percent_area}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
