% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-printcolordf.R
\name{make_color_contrast}
\alias{make_color_contrast}
\title{Make color contrast}
\usage{
make_color_contrast(
  x,
  y = NULL,
  bg = "white",
  L_threshold = 65,
  C_floor = 90,
  L_hi = 95,
  L_lo = 40,
  blend_preset = "ryb",
  do_plot = FALSE,
  cex = 2,
  ...
)
}
\arguments{
\item{blend_preset}{\code{character} string passed as \code{preset} to
\code{colorjam::blend_colors()} to define the color wheel used
during color blending operations.}
}
\description{
Make color contrast
}
\details{
This function provides a simple method to ensure a color
has adequate visual contrast with a background color,
while retaining some color saturation. For example,
red on red background will return something close to
pink, so the pink retains the red color saturation
but is visually distinct from the background color red.

Similarly, pink on pink will return something close to
red.
}
\examples{
x <- c("firebrick", "dodgerblue", "gold", "pink", "white");
y <- rep("red3", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- rep("pink1", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- rep("gold", 4);
make_color_contrast(x, y, do_plot=TRUE);

y <- c("red4", "aquamarine4", "blue3", "yellow", "pink2")
make_color_contrast(x, y, do_plot=TRUE, C_floor=140);

}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
