% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-overlap.R
\name{make_venn_combn_df}
\alias{make_venn_combn_df}
\title{Make full Venn combination data.frame}
\usage{
make_venn_combn_df(x, include_zero = FALSE, sep = "&", ...)
}
\arguments{
\item{x}{either character vector of set names, or integer number
of sets.}

\item{include_zero}{logical indicating whether to include one
row with all zeros, in the event of counting factors where
some factor levels may not be present in any sets.}

\item{sep}{character separator used when combining set names into
a single label.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} where rownames indicate each possible
Venn overlap, colnames indicate each set name, and values
are \code{0} or \code{1} indicating whether each set should have
a value in each row.
}
\description{
Make full Venn combination data.frame
}
\details{
This function returns a \code{data.frame} with all possible
combinations of set overlaps for the number of sets
provided.
}
\examples{
make_venn_combn_df(3);
make_venn_combn_df(3, include_zero=TRUE);

make_venn_combn_df(letters[1:3]);

}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{assign_degree_groups}()},
\code{\link{check_systemfonts_family}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_degrees_clockwise}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{render_venndir_footnotes}()},
\code{\link{shrink_df}()},
\code{\link{simple_ellipse}()},
\code{\link{subset_systemfonts}()},
\code{\link{venndir_label_style}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
