% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-overlap.R
\name{make_venn_combn_df}
\alias{make_venn_combn_df}
\title{Make full Venn combination data.frame}
\usage{
make_venn_combn_df(x, include_zero = FALSE, sep = "&", ...)
}
\arguments{
\item{x}{either character vector of set names, or integer number
of sets.}

\item{include_zero}{logical indicating whether to include one
row with all zeros, in the event of counting factors where
some factor levels may not be present in any sets.}

\item{sep}{character separator used when combining set names into
a single label.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} where rownames indicate each possible
Venn overlap, colnames indicate each set name, and values
are \code{0} or \code{1} indicating whether each set should have
a value in each row.
}
\description{
Make full Venn combination data.frame
}
\details{
This function returns a \code{data.frame} with all possible
combinations of set overlaps for the number of sets
provided.
}
\examples{
make_venn_combn_df(3);
make_venn_combn_df(3, include_zero=TRUE);

make_venn_combn_df(letters[1:3]);

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_venndir_label}()},
\code{\link{print_color_df}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{shrink_df}()},
\code{\link{sp_percent_area}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
