% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-test.R
\name{make_venn_test}
\alias{make_venn_test}
\title{Generate test sets for venndir}
\usage{
make_venn_test(
  n_items = 1e+06,
  n_sets = 4,
  do_signed = FALSE,
  concordance = 0.5,
  min_size = ceiling(n_items/50),
  max_size = ceiling(n_items/2),
  items = NULL,
  sizes = NULL,
  seed = 123,
  item_prefix = "item_",
  ...
)
}
\arguments{
\item{n_items}{\code{integer} total number of items available
to all sets, also known as the universe size.}

\item{n_sets}{\code{integer} number of sets that contain items.}

\item{do_signed}{\code{logical} indicating whether to return signed
sets, which indicate directionality with \code{-1} or \code{1} values,
named by the items.}

\item{concordance}{\code{numeric} between -1 and 1, used when \code{do_signed=TRUE}.
This value imposes an approximate amount of concordance between
random pairs of sets, using the concordance equation:
\code{concordance = (agree - disagree) / (agree + disgree)} where
\code{(agree + disagree) = n}. This equation approximates
the number of items that agree as:
\code{agree = ceiling((concordance * n + n) / 2)}.}

\item{min_size}{\code{integer} minimum range of items that may
be contained in each set.}

\item{max_size}{\code{integer} maximum range of items that may
be contained in each set.}

\item{items}{\code{vector} or \code{NULL} that contains the universe
of items. When \code{items} is defined, \code{n_items} is ignored.}

\item{sizes}{\code{vector} of \code{integer} values, or \code{NULL}, indicating
the size of each set. When \code{sizes} is defined, \code{min_size} and
\code{max_size} is ignored. When \code{sizes} is defined, \code{names(sizes)}
are used as names for each set.}

\item{seed}{\code{numeric} or \code{NULL} used with \code{set.seed()} for data
reproducibility. When \code{seed=NULL} then \code{set.seed()} is not called.}

\item{item_prefix}{\code{character} string used as prefix for item names,
default \code{"item_"}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of items, either as a list of item vectors,
or when \code{do_signed=TRUE} the list of vectors, where vector
names contain the items, and vector values are signed values
from \code{c(-1, 1)}.
}
\description{
Generate test sets for venndir
}
\details{
This function generates data to use as test input to
Venn diagram functions. It can generate sets of items,
or signed sets (integer values \code{-1}, \code{1}) named by
item.

This function defines a range of set sizes, using \code{min_size}
and \code{max_size}, with roughly square-root sequence of sizes
between these two extremes.

Note that the universe size represents the total available
items, but not necessarily the total number of items
represented by the sets. For example, if \code{n_items=1000000},
\code{max_size=500} and \code{n_sets=3} then the maximum number of
items actually represented is \code{1500}.

The universe can be defined using optional argument \code{items},
which takes priority over \code{n_items}.

The specific size of each set can be defined with optional
argument \code{sizes}, which takes priority over \code{min_size}, and
\code{max_size}.
}
\examples{
## basic setlist without signed direction
setlist <- make_venn_test(n_items=100,
   n_sets=3,
   min_size=5,
   max_size=25)
set_im <- list2im_opt(setlist);
table(jamba::pasteByRow(as.matrix(set_im)*1))

## basic setlist with signed direction
setlist <- make_venn_test(n_items=100,
   n_sets=3,
   do_signed=TRUE)
jamba::sdim(setlist);

## some example overlap summaries
sv1 <- signed_overlaps(setlist=setlist, "overlap")
sv1

## Familiar named overlap counts
jamba::nameVector(sv1[,c("count","sets")])

## directional count table for each combination
sv2 <- signed_overlaps(setlist=setlist, "each")
sv2

## directional count table for agreement or mixed
sv3 <- signed_overlaps(setlist=setlist, "agreement")
sv3

## signed incidence matrix
imv <- list2im_value(setlist)
dim(imv)
head(imv)

## text venn diagram
textvenn(setlist, overlap_type="overlap")

## text venn diagram with signed direction
textvenn(setlist, overlap_type="each")

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{expand_range}()},
\code{\link{get_venndir_curate_df}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{modify_venndir_overlap}()},
\code{\link{nudge_venndir_label}()},
\code{\link{plot,Venndir,ANY-method}},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{three_point_angle}()},
\code{\link{venndir_legender}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir utility}
