% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-angles.R
\name{mean_degree_arc}
\alias{mean_degree_arc}
\title{Mean arc angle in degrees}
\usage{
mean_degree_arc(x, use_range = FALSE, use_median = FALSE, do_plot = FALSE, ...)
}
\arguments{
\item{x}{\code{numeric} vector of angles in degrees.}

\item{use_range}{\code{logical} indicating whether to return the
mean degree angle across the range, which therefore only
uses the first and last angular values.}

\item{use_median}{\code{logical} indicating whether to use \code{median()}
instead of \code{mean()}, included here for convenience.}

\item{...}{additional arguments are ignored.}
}
\description{
Mean arc angle in degrees
}
\details{
This function differs from \code{mean_degrees()} in that it finds
the mean angle in degrees from angles along an arc, guaranteeing
that the mean angle is along that numeric arc. It is intended
that the arc does not cover more than 360 degrees, and for angles
whose numeric values are increasing.
}
\examples{
set.seed(1);
steps <- sample((1:12)^1.5, size=14, replace=TRUE);
steps <- sort(steps);
x <- cumsum(steps);
x;

opar <- par("mfrow"=c(2, 3));
on.exit(par(opar));
mean_degree_arc(x, do_plot=TRUE);
mean_degree_arc(x, use_median=TRUE, do_plot=TRUE);
mean_degree_arc(x, use_range=TRUE, do_plot=TRUE);

x <- x + 235;
mean_degree_arc(x, do_plot=TRUE);
mean_degree_arc(x, use_median=TRUE, do_plot=TRUE);
mean_degree_arc(x, use_range=TRUE, do_plot=TRUE);

}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{mean_degrees}()},
\code{\link{rescale_coordinates}()},
\code{\link{spread_degrees}()}
}
\concept{venndir spatial}
