% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{nudge_polygon_coords}
\alias{nudge_polygon_coords}
\title{Nudge polygon coordinates}
\usage{
nudge_polygon_coords(polygon_list, nudge, ...)
}
\description{
Nudge polygon coordinates
}
\details{
This function differs from \code{nudge_polygon_list()} in that all polygons
are nudged the exact same amount. If there are nested polygons, they
are iteratively all nudged the same.
}
\examples{
D <- list(
   x=c(-3, 3, 3, 0, -3),
   y=c(-3, -3, 1.5, 4, 1.5))
E <- list(
   x=c(-3, 3, 3, -3),
   y=c(-3, -3, 3, 3))
DElist <- list(D=D, E=E, DE=list(D=D, E=E))
nudge <- c(x=10, y=-10)
new_polygon_list <- nudge_polygon_coords(polygon_list=DElist, nudge=nudge)
plot_polygon_list(new_polygon_list)

}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
