% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{nudge_sp}
\alias{nudge_sp}
\title{Nudge SpatialPolygons}
\usage{
nudge_sp(sp, sp_nudge = NULL, rotate_degrees = 0, ...)
}
\arguments{
\item{sp}{\code{sp::SpatialPolygons} object}

\item{sp_nudge}{\code{list} whose names are found in \code{names(sp)},
and whose values are \code{x} and \code{y} coordinates to be moved.}

\item{rotate_degrees}{\code{numeric} value in degrees (0, 360) to
rotate the \code{sp} object and all contained polygons. (Not yet
implemented.)}

\item{...}{additional arguments are ignored.}
}
\description{
Nudge SpatialPolygons
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{make_venn_test}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()}
}
\concept{venndir utility}
