% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{nudge_sp}
\alias{nudge_sp}
\title{Nudge SpatialPolygons}
\usage{
nudge_sp(sp = NULL, sp_nudge = NULL, ...)
}
\arguments{
\item{sp}{\code{sp::SpatialPolygons} object}

\item{sp_nudge}{\code{list} whose names are found in \code{names(sp)},
and whose values are \code{x} and \code{y} coordinates to be moved.}

\item{...}{additional arguments are ignored.}

\item{rotate_degrees}{\code{numeric} value in degrees (0, 360) to
rotate the \code{sp} object and all contained polygons. (Not yet
implemented.)}
}
\value{
object \code{sp::SpatialPolygons}.
}
\description{
Nudge SpatialPolygons
}
\details{
This helper function is intended to take \code{sp::SpatialPolygons}
and "nudge" (move by adding a scalar value to each coordinate)
only a subset of polygons identified by name.

It differs from \code{rescale_sp()} because \code{rescale_sp()} manipulates
all contained polygons together.
}
\examples{
sp <- sp_ellipses(c(3, 2), c(2, 3),
   xradius=c(1, 4),
   yradius=c(5, 2))
opar <- par("mfrow"=c(1,2));
on.exit(par(opar));
plot(sp, col=c("#FF000077", "#FFDD0077"),
   xlim=c(-2, 9));
axis(1, las=2); axis(2, las=2);
points(x=c(3, 2), y=c(2, 3), pch=c("1", "2"), add=TRUE);

sp2 <- nudge_sp(sp,
   sp_nudge=list("2"=c(3, -2))
)
plot(sp2, col=c("#FF000077", "#FFDD0077"),
   xlim=c(-2, 9));
plot(sp[2], border="blue", lty="dotted", add=TRUE);
axis(1, las=2); axis(2, las=2);
arrows(x0=2, x1=5, y0=3, y1=1)
points(x=c(3, 5), y=c(2, 1), pch=c("1", "2"), add=TRUE);

}
\seealso{
Other venndir spatial: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{intersect_polygons}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_sp}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()}
}
\concept{venndir spatial}
