% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-nudge-label.R
\name{nudge_venndir_label}
\alias{nudge_venndir_label}
\title{Nudge venndir label}
\usage{
nudge_venndir_label(
  venndir_output,
  set = NULL,
  x_offset = 0,
  y_offset = 0,
  unit_type = c("relative", "absolute"),
  label_types = c("main", "signed"),
  label_location = c("all", "inside", "outside"),
  show_label = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{venndir_output}{output from \code{venndir()} as \code{Venndir} object.}

\item{set}{\code{character} name of the set or overlap to adjust}
}
\value{
\code{Venndir} object as output from \code{venndir()}
}
\description{
Nudge venndir label, work in progress currently
}
\examples{
setlist1 <- make_venn_test(100, 3, do_signed=TRUE)

vo1 <- venndir(setlist1,
   overlap_type="each",
   return_items=TRUE,
   label_style="lite_box",
   main="Default venndir")
render_venndir(vo1)

vo2 <- nudge_venndir_label(vo1,
   set=c("set_A&set_B&set_C"),
   x_offset=0.05,
   y_offset=-0.05)
render_venndir(vo2)

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{expand_range}()},
\code{\link{get_venn_shapes}()},
\code{\link{gridtext_make_outer_box}()},
\code{\link{gridtext_richtext_grob}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{three_point_angle}()},
\code{\link{venndir_legender}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir utility}
