% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-conversion.R
\name{overlaplist2setlist}
\alias{overlaplist2setlist}
\title{venndir conversion from overlap list to setlist}
\usage{
overlaplist2setlist(x, sep = "&", ...)
}
\value{
\code{list} where the list names are the names of each set,
and values of each list element is a vector of items.
}
\description{
venndir conversion from overlap list to setlist
}
\details{
This function takes input in the form of named \code{list}
of \code{vectors}. The list names represent set overlaps.
The vectors each contain items contained in that
overlap. This content is similar to the input to
\code{counts2setlist()} except in that case the input
only contains the number of items per overlap,
and not the items. In this case each overlap contains
the vector of items also.
}
\examples{
# example from eulerr::euler()
av_overlap_list <- list(
   AV="BATMAN<br>VILLAIN",
   ML="MOUNTAIN<br>LION",
   T="TODDLER",
   `AV&ML`="Wants to<br>kill you",
   `AV&T`="Goes off on<br>a lot of rambling<br>monologues",
   `ML&T`="Resists<br>taking a bath",
   `AV&ML&T`="Impossible<br>to reason with"
)
setlist <- overlaplist2setlist(av_overlap_list);

# Venn diagram
venndir(setlist,
   show_items="item",
   item_angle=0,
   item_cex=rep(c(2, 1.6, 1.35, 1.5), c(3, 1, 2, 1)))

# Proportional Venn (Euler) diagram
venndir(setlist, proportional=TRUE)

}
\seealso{
Other venndir conversion: 
\code{\link{counts2setlist}()},
\code{\link{im2list}()},
\code{\link{im_value2list}()},
\code{\link{list2im_opt}()},
\code{\link{list2im_value}()}
}
\concept{venndir conversion}
