% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{plot_polygon_list}
\alias{plot_polygon_list}
\title{Plot polygon_list using base R}
\usage{
plot_polygon_list(
  polygon_list,
  col = NULL,
  border = "black",
  lwd = 1,
  add = FALSE,
  asp = 1,
  bty = "n",
  xaxt = "n",
  yaxt = "n",
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  rule = c("evenodd", "none"),
  ...
)
}
\description{
Plot polygon_list using base R
}
\examples{
counts <- c(A=1, B=2, `A&B`=3, C=5, `B&C`=2, `A&C`=2, `A&B&C`=1)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
plot_polygon_list(polygon_list,
   col=colorjam::rainbowJam(length(polygon_list), alpha=0.5))

polygon_list2 <- list(A=polygon_list$A, BC=polygon_list[c("B", "C")])
plot_polygon_list(polygon_list2,
   col=colorjam::rainbowJam(length(polygon_list2), alpha=0.5))

}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()},
\code{\link{xy_list_to_polygon_list}()}
}
\concept{venndir polygons}
