% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{polygon_ellipses}
\alias{polygon_ellipses}
\title{Make polygon_list ellipses}
\usage{
polygon_ellipses(
  xcenter,
  ycenter,
  setnames = NULL,
  xradius = 1,
  yradius = 2,
  rotation_degrees = c(0),
  n = 60,
  ...
)
}
\arguments{
\item{xcenter, ycenter}{\code{numeric} vector that defines the x and y
coordinate position of the center of each ellipse.}

\item{setnames}{\code{vector} that contains names for each circle, stored
as \code{names()} for each polygon.
When \code{setnames} is \code{NULL} then index numbers are used.}

\item{xradius, yradius}{\code{numeric} vector that defines the radius
of each ellipse along the x-axis and y-axis, respectively.
Each \code{vector} is recycled to \code{length(xcenter)}.}

\item{n}{\code{integer} value indicating the number of subdivisions to
use in the circle.}

\item{...}{additional arguments are ignored.}

\item{rotation_degree}{\code{numeric} vector representing degrees to
rotate each ellipse after it is created, where values are
conformed to between \code{0} and \code{360}, rotating clockwise.}
}
\value{
\code{JamPolygon} object
}
\description{
Make polygon_list ellipses
}
\details{
This function creates one or more ellipses as polygon_list \code{list} objects.
}
\examples{
ejp <- polygon_ellipses(c(3, 2), c(2, 3),
   xradius=c(1, 4),
   yradius=c(5, 2))
plot(ejp, fill=c("#FF000077", "#FFDD0077"))

}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{[,JamPolygon,ANY,ANY,ANY-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
