% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{polygon_label_outside}
\alias{polygon_label_outside}
\title{Polygon label outside}
\usage{
polygon_label_outside(
  sp,
  which_sp = NULL,
  center = NULL,
  distance = 1,
  center_method = c("label", "bbox"),
  vector_method = c("farthest", "label"),
  segment_type = c("vector", "nearest"),
  min_degrees = 15,
  debug = FALSE,
  seed = 1,
  ...
)
}
\arguments{
\item{sp}{\code{sp::SpatialPolygonsDataFrame} or \code{sp::SpatialPolygons}
object, which may contain multiple polygons.}

\item{which_sp}{\code{integer} or \code{NULL}; when \code{which_sp} contains one
or more \code{integer} values, they refer to polygons in \code{sp}, and
each will be analyzed in sequence. When \code{which_sp=NULL} then
all the polygons in \code{sp} will be analyzed in sequence.}

\item{center}{\code{numeric} vector or matrix with two values indicating
the center position. When \code{center=NULL} then the center is
determined using a method defined by \code{center_method}.}

\item{distance}{\code{numeric} value indicating the absolute coordinate
distance away from the perimiter of \code{sp} to place labels.
This value is effectively the same as buffer used in
\code{rgeos::gBuffer()}.}

\item{center_method}{\code{character} string indicating the method to
determine the \code{center}: \code{"label"} uses the mean x,y coordinate
of all the polygon label positions; \code{"bbox"} uses the mean x,y
coordinate of the bounding box that encompasses \code{sp}. The effect
is to extend outer labels radially around this center point.
Using the mean label position with \code{center_method="label"}
is helpful because it ensures labels are extended in all directions
even when most labels are in one upper or lower section of
the \code{sp} polygons.}

\item{vector_method}{\code{character} string indicating the point to draw
a vector from the \code{center} position for each polygon:
\code{"farthest"} chooses the farthest point in each polygon from
the \code{center} point; \code{"label"} uses the position of the label.
Using \code{vector_method="farthest"} works well to label the outer
edge of each polygon. Using \code{vector_method="label"} works well
when the polygon label is set based upon the interior open area
of the polygon.}

\item{segment_type}{\code{character} string indicating how to draw the
line segment from the outside label back to the polygon:
\code{"vector"} defines the line segment by extending the line
used by \code{vector_method} along the same direction; \code{"nearest"}
defines a new line to the nearest point on the polygon.}

\item{min_degrees}{\code{numeric} value passed to \code{spread_degrees()}
with the minimum angle in degrees for labels to be placed
around the outside of a polygon.}

\item{debug}{\code{logical} indicating whether to plot the results
using base R graphics. When \code{debug=2} it will also add some
visual decorations showing the center point, the points
determined by polygon boundaries.}

\item{...}{additional arguments are ignored.}
}
\description{
Polygon label outside
}
\details{
This function takes a \code{sp::SpatialPolygons} object and
determines an appropriate place for a label outside the
polygon, with line segment drawn back to the outer border.

The expected input is \code{sp::SpatialPolygonsDataFrame} which
contains multiple polygons, and one of the polygons should
be labeled.
}
\examples{
setlist <- list(A=letters, B=sample(letters, 4));
C <- sample(setdiff(letters, setlist$B), 4);
setlist$C <- C;

vo <- venndir(setlist, proportional=FALSE, expand_fraction=0.2);
segment1 <- polygon_label_outside(sp=vo$venn_spdf)
lines(segment1[[1]], col=vo$venn_spdf$border[1], lwd=2)
jamba::drawLabels(txt=vo$venn_spdf$label[1],
   labelCex=1.5,
   x=segment1[[1]][2,1], y=segment1[[1]][2,2])

i <- 5;
lines(segment1[[i]], col=vo$venn_spdf$border[i], lwd=2)
jamba::drawLabels(txt=vo$venn_spdf$label[i],
   labelCex=1.5,
   x=segment1[[i]][2,1], y=segment1[[i]][2,2])

par("xpd"=NA);
par("mfrow"=c(4,2));
st <- "nearest";
for (cm in c("label", "bbox")) {
for (vm in c("farthest", "label")) {
   for (st in c("vector", "nearest")) {
      vo <- venndir(setlist, proportional=FALSE,
         expand_fraction=0.2);
      title(cex.main=1,
         main=paste0("vector_method:", vm,
            "\ncenter_method:", cm,
            "\nsegment_type:", st));
      segmentxy <- polygon_label_outside(sp=vo$venn_spdf,
         center_method=cm,
         vector_method=vm,
         debug=2)
   }
}
}

par("mfrow"=c(1, 2));
st <- "nearest";
for (cm in c("label", "bbox")) {
for (vm in c("farthest", "label")) {
   for (st in c("vector", "nearest")) {
      set.seed(12);
      vo <- venndir(setlist, proportional=TRUE,
         expand_fraction=0.2);
      title(cex.main=1,
         main=paste0("vector_method:", vm,
            "\ncenter_method:", cm,
            "\nsegment_type:", st));
      segmentxy <- polygon_label_outside(sp=vo$venn_spdf,
         center_method=cm,
         vector_method=vm,
         debug=2)
   }
}
}

vo <- venndir(setlist, proportional=TRUE);
segmentxy <- polygon_label_outside(sp=vo$venn_spdf, debug=2)

set.seed(1)
vo <- venndir(setlist, proportional=TRUE, expand_fraction=0.3);
segmentxy <- polygon_label_outside(sp=vo$venn_spdf,
   debug=TRUE, segment_type="nearest")
segmentxy <- polygon_label_outside(sp=vo$venn_spdf, 
   debug=TRUE, segment_type="vector")

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
