% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-printcolordf.R
\name{print_color_df}
\alias{print_color_df}
\title{Print colorized text table}
\usage{
print_color_df(
  df,
  dfcolor = NULL,
  dfinvert = NULL,
  dfjustify = "centre",
  header = FALSE,
  padding = 2,
  pad_type = 1,
  comment = FALSE,
  timeStamp = FALSE,
  ...
)
}
\arguments{
\item{df}{\code{data.frame} with content to be printed to the R console.}

\item{dfcolor}{\code{data.frame} with R compatible colors in each cell.}

\item{dfinvert}{\code{data.frame} with \code{logical} values indicating whether
each color should be inverted \code{TRUE} so the cell color is applied
as a background color. In this case the foreground text color is
determined using \code{jamba::setTextContrastColor()}.}

\item{dfjustify}{\code{character} or \code{data.frame} passed to \code{format()}
to determine the text alignment in each cell. When supplied as
a vector, it describes each column of output; when supplied as
a \code{data.frame} it describes each cell.}

\item{header}{\code{logical} indicating whether to print column names
at the top of the output.}

\item{padding}{\code{integer} value that defines the number of characters
added before and after each text term to add visual space between
adjacent cells.}

\item{comment, timeStamp}{\code{logical} arguments passed to
\code{jamba::printDebug()}, which indicate whether to print output
as a comment (with \code{"#"} at the start of each line), and whether
to include a time stamp. The default \code{FALSE} turn these options
off.}

\item{...}{additional arguments are ignored.}
}
\description{
Print colorized text table
}
\details{
This utility function takes \code{df} as a \code{data.frame} input,
and prints colorized text output.

Colors are defined using \code{dfcolor} provided as a \code{data.frame}.

Colors can be inverted using \code{dfinvert} that contains logical
\code{TRUE} or \code{FALSE} values, provided as a \code{data.frame}.
}
\examples{
set.seed(2);
df <- matrix(sample(head(colors(), 151), 12), ncol=3);
dfinvert <- matrix(sample(c(FALSE, FALSE, TRUE), 12, replace=TRUE), ncol=3);
print_color_df(df, df, dfinvert);

print_color_df(df, df, dfinvert, padding=1);

colnames(df) <- c("left", "centre", "right");
print_color_df(df, df, dfinvert,
   header=TRUE,
   padding=4,
   dfjustify=c("left", "centre", "right"));

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_segment}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
