% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-render-jp.R
\name{render_venndir}
\alias{render_venndir}
\title{Render venndir output}
\usage{
render_venndir(
  venndir_output = NULL,
  expand_fraction = 0,
  font_cex = 1,
  item_cex = NULL,
  item_cex_factor = 4,
  plot_warning = TRUE,
  show_labels = NULL,
  show_items = c(NA, "none", "sign item", "item", "sign"),
  item_degrees = 0,
  max_items = 100,
  show_zero = TRUE,
  show_segments = TRUE,
  segment_buffer = -0.1,
  label_preset = c("none"),
  label_style = c("custom", "basic", "fill", "shaded", "shaded_box", "lite", "lite_box"),
  fontfamily = "Arial",
  inside_percent_threshold = 0,
  item_style = c("default", "text", "gridtext"),
  item_buffer = -0.15,
  group_labels = TRUE,
  adjust_center = FALSE,
  draw_legend = TRUE,
  legend_x = "bottomright",
  legend_font_cex = 1,
  show_label = NA,
  display_counts = TRUE,
  draw_buffer = FALSE,
  ...
)
}
\arguments{
\item{venndir_output}{\code{Venndir} output from \code{venndir()}, or
\code{list} with element \code{"vo"} as a \code{Venndir} object.}

\item{expand_fraction}{\code{numeric} value indicating how much to
expand the figure range beyond the default calculated for
the Venn diagram. Values above zero cause the Venn diagram
to be slighly smaller.}

\item{font_cex}{\code{numeric} scalar to adjust font sizes.}

\item{item_cex}{\code{numeric} scalar applied to item labels in each overlap
in order. When \code{length(item_cex) == 1} it is applied uniformly
across all overlaps, otherwise it is recycled to the total
number of overlaps.
When provided, it is used instead of any adjustments to item label
sizes based upon proportional area of each overlap.}

\item{item_cex_factor}{\code{numeric} value used to adjust pre-calculated
item fontsizes. This value is used to adjust the item label sizes,
which may also be adjusted proportional to the area of each overlap,
in which case \code{item_cex_factor} is used to adjust those relative
label sizes.}

\item{plot_warning}{\code{logical} indicating whether to include a warning
when one or more non-zero overlap counts cannot be displayed
in the figure.
\strong{Not yet re-implemented since version 0.0.30.900.}}

\item{show_labels}{\code{character} string to define the labels to display,
and where they should be displayed.
The definition uses a single letter to indicate each type of label
to display, using UPPERCASE to display the label outside the Venn shape,
and lowercase to display the label inside the Venn shape.
The default \code{"Ncs"} displays _N_ame (outside), _c_ount (inside),
and _s_igned count (inside).

The label types are defined below:
\itemize{
\item _N_ame: "n" or "N" - the set name, by default it is displayed.
\item _O_verlap: "o" or "O" - the overlap name, by default it is hidden,
because these labels can be very long, also the overlap should be
evident in the Venn diagram already.
\item _c_ount: "c" or "C" - overlap count, independent of the sign
\item _p_ercentage: "p" or "P" - overlap percentage, by default hidden,
but available as an option
\item _s_igned count: "s" or "S" - the signed overlap count, tabulated
based upon \code{overlap_type} ("each", "concordant", "agreement", etc/)
\item _i_tems: "i" only, by default hidden. When enabled, item labels
defined by \code{show_items} are spread across the specific Venn overlap
region.
}}

\item{show_items}{\code{character} used to define the item label,
only used when the \code{show_label} entry includes \code{"i"} which
enables item display inside the Venn diagram.
\itemize{
\item \code{"item"}: shows only the item labels
\item \code{"sign"}: shows only the sign of each item
\item \code{"sign items"}: shows the sign and item together
(or \code{"item sign"} will show the item, then the sign).
}}

\item{item_degrees}{\code{numeric} angle (default 0) in degrees used
to adjust item label display.}

\item{max_items}{\code{numeric} (default 3000) indicating the maximum number
of item labels to display when enabled.}

\item{show_zero}{\code{logical} (default FALSE) indicating whether empty
overlaps are labeled with count zero \code{0}. When \code{show_zero=TRUE} the
count zero label is displayed, otherwise no count label is shown.}

\item{show_segments}{\code{logical} (default TRUE) indicating whether to
draw a line segment from the Venn region to any label
positioned outside.}

\item{segment_buffer}{\code{numeric} (default -0.1) indicating the depth
inside each Venn region a line segment will be drawn, relevant
only when \code{show_segments=TRUE}.}

\item{label_preset}{\code{character} deprecated in favor of \code{show_labels}.
This argument is passed to \code{venndir_label_style()}.}

\item{label_style}{\code{character} string indicating the style for labels.
Label color is adjusted based upon the determined background color,
determined based upon the label fill color, and either the
device background color, or Venn overlap fill color.
There are pre-defined label styles.
\itemize{
\item \code{"basic"} no background shading
\item \code{"fill"} an opaque colored background
\item \code{"shaded"} a partially transparent colored background
\item \code{"lite"} a partially transparent lite background
\item \code{"box"} adds a dark border around the label region
}}

\item{fontfamily}{\code{character} font family}

\item{inside_percent_threshold}{\code{numeric} (default 0) indicating the
percent area that a Venn overlap region must contain in order
for the count label to be displayed inside the region,
otherwise the label is displayed outside the region.
Values are expected to range from 0 to 100.}

\item{item_style}{\code{character} string (default "text") indicating
the style to display item labels when they are enabled.
\itemize{
\item \code{"default"} detects whether item labels contain \code{"<br>"} for newlines,
and uses \code{"gridtext"} if that is the case, otherwise it uses \code{"text"}
which is markedly faster.
\item \code{"text"} option is substantially faster, but does not allow
markdown.
\item \code{"gridtext"}:  substantially slower for a large number of labels,
but enables use of limited markdown by calling
\code{gridtext::richtext_grob()}.
Mostly useful for \code{venn_meme()}.
}}

\item{item_buffer}{\code{numeric} value (default -0.15) indicating the buffer
adjustment applied to Venn overlap regions before arranging item
labels. Passed to \code{label_fill_JamPolygon()} via \code{render_venndir()}.
Negative values are recommended, so they shrink the region.}

\item{group_labels}{\code{logical} (default TRUE) indicating whether to group
label components together, therefore drawing fill and border
around the group instead of each component. In most cases this
setting should be TRUE.}

\item{adjust_center}{\code{logical} (default TRUE) used when labels are grouped,
whether the group should be re-centered on the target point.
Try \code{adjust_center=FALSE} if wide label groups are adjusted
so that the count label is too far left.}

\item{draw_legend}{\code{logical} (default TRUE) indicating whether to draw
a legend, calling \code{venndir_legender()}.}

\item{legend_x}{\code{character} passed to \code{venndir_legender()} to customize
the position of the legend.}

\item{legend_font_cex}{\code{numeric} scalar to adjust the legend font size.}

\item{...}{additional arguments are passed to \code{render_venndir()}.}
}
\description{
Render venndir output
}
\seealso{
Other venndir core: 
\code{\link{signed_overlaps}()},
\code{\link{textvenn}()},
\code{\link{venn_meme}()},
\code{\link{venndir}()}
}
\concept{venndir core}
