% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{rescale_sp}
\alias{rescale_sp}
\title{Rescale a SpatialPolygons object}
\usage{
rescale_sp(
  sp,
  rotate_degrees = 0,
  scale = c(1, 1),
  shift = c(0, 0),
  center = NULL,
  share_center = FALSE,
  update_bbox = TRUE,
  ...
)
}
\arguments{
\item{rotate_degrees}{\code{numeric} value in degrees indicating
rotation around the \code{center}, where positive values are
clockwise rotation. This rotation is only applied to two
columns in \code{x} defined by \code{rotation_axes}.}

\item{scale}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. After subtracting the \code{center}, the coordinates
in each column are multiplied by the \code{scale}.}

\item{shift}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}. The coordinates in each column are added to
the \code{shift}, after applying \code{scale} and \code{rotate_degrees}}

\item{center}{\code{numeric} vector whose values are expanded to length
\code{ncol(x)}, indicating the center point used for \code{scale} and
\code{rotate_degrees} transformations. When \code{center=NULL} it
is derived from the bounding box, which is the mean of the range
for each column in \code{x}.}

\item{share_center}{\code{logical} indicating whether all polygons
should share the same center, where \code{share_center=TRUE} will
adjust everything collectively, and \code{share_center=FALSE} will
adjust each polygon independently relative to its own center
coordinate.}

\item{...}{additional arguments are ignored.}

\item{p}{object \code{sp::SpatialPolygons}}
}
\value{
object \code{sp::SpatialPolygons}
}
\description{
Rescale a SpatialPolygons object
}
\details{
This function simply applies \code{rescale_coordinates()} to an
object \code{sp::SpatialPolygons}, and it does so by calling
\code{rescale_ps} on each \code{sp::Polygons} object contained in
the \code{sp} input.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
