% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-shrinkdf.R
\name{shrink_df}
\alias{shrink_df}
\title{Shrink data.frame by group}
\usage{
shrink_df(
  df,
  by,
  string_func = jamba::cPasteU,
  num_func = mean,
  extra_funcs = NULL,
  do_test = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Shrink data.frame by group
}
\details{
This function condenses a `data.frame` by groups of rows,
applying an appropriate function to `character` columns,
and `numeric` columns. It is intended to be a simple but
configurable tool for the majority of scenarios.

This function uses `data.table` for overall speed.
}
