% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-shrinkdf.R
\name{shrink_df}
\alias{shrink_df}
\title{Shrink data.frame by group}
\usage{
shrink_df(
  df,
  by,
  string_func = jamba::cPasteU,
  num_func = mean,
  extra_funcs = NULL,
  do_test = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Shrink data.frame by group
}
\details{
This function condenses a \code{data.frame} by groups of rows,
applying an appropriate function to \code{character} columns,
and \code{numeric} columns. It is intended to be a simple but
configurable tool for the majority of scenarios.

This function uses \code{data.table} for overall speed.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
