% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-shrinkdf.R
\name{shrink_df}
\alias{shrink_df}
\title{Shrink data.frame by group}
\usage{
shrink_df(
  df,
  by,
  string_func = jamba::cPasteU,
  num_func = mean,
  extra_funcs = NULL,
  do_test = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Shrink data.frame by group
}
\details{
This function condenses a \code{data.frame} by groups of rows,
applying an appropriate function to \code{character} columns,
and \code{numeric} columns. It is intended to be a simple but
configurable tool for the majority of scenarios.

This function uses \code{data.table} for overall speed.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{make_venn_test}()},
\code{\link{nudge_sp}()},
\code{\link{print_color_df}()}
}
\concept{venndir utility}
