% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-shrinkdf.R
\name{shrink_df}
\alias{shrink_df}
\title{Shrink data.frame by group}
\usage{
shrink_df(
  df,
  by,
  string_func = jamba::cPasteU,
  num_func = mean,
  extra_funcs = NULL,
  do_test = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Shrink data.frame by group
}
\details{
This function condenses a \code{data.frame} by groups of rows,
applying an appropriate function to \code{character} columns,
and \code{numeric} columns. It is intended to be a simple but
configurable tool for the majority of scenarios.

This function uses \code{data.table} for overall speed.
}
\examples{
testdf <- data.frame(check.names=FALSE,
   SYMBOL=rep(c("ACTB", "GAPDH", "PPIA"), c(2, 3, 1)),
   `logFC B-A`=c(1.4, 1.4, 2.3, NA, 2.3, 5.1),
   probe=paste0("probe", 1:6))
shrink_df(testdf, by="SYMBOL", num_func=function(x){mean(x, na.rm=TRUE)})

# 60,000 row simulation
testdftall <- do.call(rbind, lapply(1:10000, function(i){
   idf <- testdf;
   idf$SYMBOL <- paste0(idf$SYMBOL, "_", i);
   idf;
}))
head(testdftall, 12)
shrunk_tall <- shrink_df(testdftall,
   by="SYMBOL",
   num_func=function(x){mean(x, na.rm=TRUE)})

if (requireNamespace("jamses", quietly=TRUE)) {
   shrunk_tall2 <- jamses::shrinkDataFrame(testdftall,
      groupBy="SYMBOL")
   print(head(shrunk_tall2, 12))
}

}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{simple_ellipse}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
