% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-overlap.R
\name{signed_overlaps}
\alias{signed_overlaps}
\title{Signed overlaps}
\usage{
signed_overlaps(
  setlist,
  overlap_type = c("each", "overlap", "concordance", "agreement"),
  return_items = FALSE,
  return_item_labels = return_items,
  sep = "&",
  trim_label = TRUE,
  include_blanks = TRUE,
  ...
)
}
\arguments{
\item{setlist}{\code{list} of named vectors, whose names represent
set items, and whose values represent direction using
values \code{c(-1, 0, 1)}.}

\item{overlap_type}{\code{character} value indicating the type of
overlap logic: \code{"each"} records each combination of
signs; \code{"overlap"} disregards the sign and returns any match
item overlap; \code{"concordance"} represents counts for full
agreement, or \code{"mixed"} for any inconsistent overlapping
direction; \code{"agreement"} represents full agreement in direction
as \code{"agreement"}, and \code{"mixed"} for any inconsistent
direction.}

\item{return_items}{\code{logical} indicating whether to return
the items within each overlap set.}

\item{return_item_labels}{\code{logical} indicating whether to return
the directional label associated with each item. A directional
label combines the direction from \code{setlist} by item.}

\item{sep}{\code{character} used as a delimiter between set names,
the default is \code{"&"}.}

\item{trim_label}{\code{logical} indicating whether to trim the
directional label, for example instead of returning \code{"0 1 -1"}
it will return \code{"1 -1"} because the overlap name already
indicates the sets involved.}

\item{include_blanks}{\code{logical} indicating whether each set overlap
should be represented at least once even when no items are
present in the overlap. When \code{include_blanks=TRUE} is useful
in that it guarantees all possible combinations of overlaps
are represented consistently in the output.}

\item{...}{additional arguments are passed to \code{list2imsigned()}.}
}
\description{
Calculate signed, directional overlaps across sets
}
\details{
This function is the core function to summarize overlaps
that include signed directionality. It is intended for
situations where two sets may share items, but where the
signed direction associated with those items may or may
not also be shared.

One motivating example is with biological data, where
a subset of genes, proteins, or regions of genome, may be
regulated up or down, and this direction is relevant
to understanding the biological process. Two experiments
may identify similar genes, proteins, or regions of
genome, but they may not regulate them in the same
direction. This function is intended to help summarize
item overlaps alongside the directionality of each item.

The directional counts can be summarized in slightly different
ways, defined by the argument \code{overlap_type}:
\itemize{
\item \code{overlap_type="each"} - this option returns all possible
directions individually counted.
\item \code{overlap_type="concordance"} - this option returns the counts
for each consistent direction, for example \code{"up-up-up"} would
be counted, and \code{"down-down-down"} would be counted, but any
mixture of \code{"up"} and \code{"down"} would be summarized and counted
as \code{"mixed"}. For 3-way overlaps, there are 8 possible directions,
the labels are difficult to place in the Venn diagram, and are not
altogether meaningful. Note that this option is the default
for \code{venndir()}.
\item \code{overlap_type="overlap"} - this option only summarizes overlaps
without regard to direction. This option returns standard Venn
overlap counts.
\item \code{overlap_type="agreement"} - this option groups all directions
that agree and returns them as \code{"concordant"}, all others are
returned as \code{"mixed"}.
}
}
\examples{
setlist <- make_venn_test(100, 2, do_signed=TRUE);

# straight overlap counts
signed_overlaps(setlist, "overlap");

# each directional overlap count
signed_overlaps(setlist, "each");

# concordance overlap counts
signed_overlaps(setlist, "concordance");

# agreement overlap counts
signed_overlaps(setlist, "agreement");

}
\seealso{
Other venndir core: 
\code{\link{render_venndir}()},
\code{\link{textvenn}()},
\code{\link{venndir}()}
}
\concept{venndir core}
