% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-overlap.R
\name{signed_overlaps}
\alias{signed_overlaps}
\title{Signed overlaps}
\usage{
signed_overlaps(
  setlist,
  overlap_type = c("detect", "each", "overlap", "concordance", "agreement"),
  return_items = FALSE,
  return_item_labels = return_items,
  sep = "&",
  trim_label = TRUE,
  include_blanks = TRUE,
  keep_item_order = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{setlist}{\code{list} of named vectors, whose names represent
set items, and whose values represent direction using
values \code{c(-1, 0, 1)}.}

\item{overlap_type}{\code{character} value indicating the type of
overlap logic:
\itemize{
\item \code{"each"} records each combination of signs;
\item \code{"overlap"} disregards the sign and returns any match
item overlap;
\item \code{"concordance"} represents counts for full
agreement, or \code{"mixed"} for any inconsistent overlapping
direction;
\item \code{"agreement"} represents full agreement in direction
as \code{"agreement"}, and \code{"mixed"} for any inconsistent
direction.
}}

\item{return_items}{\code{logical} indicating whether to return
the items within each overlap set.}

\item{return_item_labels}{\code{logical} indicating whether to return
the directional label associated with each item. A directional
label combines the direction from \code{setlist} by item.}

\item{sep}{\code{character} used as a delimiter between set names,
the default is \code{"&"}.}

\item{trim_label}{\code{logical} indicating whether to trim the
directional label, for example instead of returning \code{"0 1 -1"}
it will return \code{"1 -1"} because the overlap name already
indicates the sets involved.}

\item{include_blanks}{\code{logical} indicating whether each set overlap
should be represented at least once even when no items are
present in the overlap. When \code{include_blanks=TRUE} is useful
in that it guarantees all possible combinations of overlaps
are represented consistently in the output.}

\item{keep_item_order}{\code{logical} default FALSE, to determine whether
items will be stored and displayed in the order they are provided.
Note: \code{keep_item_order=TRUE} enables the following behaviors:
\itemize{
\item Any \code{character} vector input will retain the order they appear.
\item Any \code{factor} vector input will sort items using factor \code{levels},
which maintains the factor level order.
\item Any named vector will use the \code{character} vector of names, keeping
the order they appear in the vector.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{list2imsigned()}.}
}
\value{
\code{data.frame} with columns intended to support \code{venndir()},
but which may be more widely useful:
\itemize{
\item \code{"sets"} - character vector with sets and overlap names.
\item one column indicating the \code{overlap_type} and corresponding values:
\itemize{
\item \code{"overlap"} - This column is always included.
\item \code{"concordance"} - includes \code{1} (concordant) and \code{-1} (discordant)
\item \code{"agreement"} - includes \code{"agreement"} and \code{"disgreement"}
\item \code{"each"} - includes sign values \code{-1} and \code{1}.
}
\item \code{"overlap"} - integer vector with overlap values, where \code{0} and \code{1}
indicate which sets contained these items. This column is always included,
even when \code{overlap_type} is something else.
\item \code{"num_sets"} - integer number of sets represented in the overlap.
\item \code{"count"} - integer number of items in the overlap.
\item one colname for each set name represented in the \code{"sets"} column,
intended to help filter by each set. Values will be \code{0} or \code{1}.
\item \code{overlap_label} - will represent only the non-0 elements from
\code{"overlap"} for convenience.
\item \code{"items"} - when \code{return_items=TRUE} this column will contain
a \code{list} (in \code{AsIs} format) of \code{character} vectors, with the items.
}
}
\description{
Calculate signed, directional overlaps across sets
}
\details{
This function is the core function to summarize overlaps
that include signed directionality. It is intended for
situations where two sets may share items, but where the
signed direction associated with those items may or may
not also be shared.

One motivating example is with biological data, where
a subset of genes, proteins, or regions of genome, may be
regulated up or down, and this direction is relevant
to understanding the biological process. Two experiments
may identify similar genes, proteins, or regions of
genome, but they may not regulate them in the same
direction. This function is intended to help summarize
item overlaps alongside the directionality of each item.

The directional counts can be summarized in slightly different
ways, defined by the argument \code{overlap_type}:
\itemize{
\item \code{overlap_type="detect"} - default behavior: each vector in \code{setlist}
is handled independently:
\itemize{
\item a vector with no names will use the vector
values as items after converting them to \code{character};
\item a named vector with \code{character} or \code{factor} values
will will use the vector names as items,
and character values as item values;
\item a named vector with \code{numeric} or \code{integer} values
will use vector names as items, and will convert
numeric values to \code{sign()}.
}
\item \code{overlap_type="each"} - this option returns all possible
directions individually counted.
\item \code{overlap_type="concordance"} - this option returns the counts
for each consistent direction, for example \code{"up-up-up"} would
be counted, and \code{"down-down-down"} would be counted, but any
mixture of \code{"up"} and \code{"down"} would be summarized and counted
as \code{"mixed"}. For 3-way overlaps, there are 8 possible directions,
the labels are difficult to place in the Venn diagram, and are not
altogether meaningful. Note that this option is the default
for \code{venndir()}.
\item \code{overlap_type="overlap"} - this option only summarizes overlaps
without regard to direction. This option returns standard Venn
overlap counts.
\item \code{overlap_type="agreement"} - this option groups all directions
that agree and returns them as \code{"concordant"}, all others are
returned as \code{"mixed"}.
}

Note that \code{overlap_type="agreement"} and \code{overlap_type="concordance"}
will not convert \code{numeric} values to \code{sign()}, so if the input
contains \code{numeric} values such as \code{1.2435} they should probably be
converted to \code{sign()} before calling \code{signed_overlaps()}, for example:
\code{signed_overlaps(lapply(setlist, sign))}
}
\examples{
setlist <- make_venn_test(100, 2, do_signed=FALSE);

# detect overlap_type
attr(signed_overlaps(setlist, "detect"), "overlap_type")

setlist <- make_venn_test(100, 2, do_signed=TRUE);

# detect overlap_type
attr(signed_overlaps(setlist, "detect"), "overlap_type")

# straight overlap counts
signed_overlaps(setlist, "overlap");

# each directional overlap count
signed_overlaps(setlist, "each");

# concordance overlap counts
signed_overlaps(setlist, "concordance");

# agreement overlap counts
signed_overlaps(setlist, "agreement");

# test to ensure factor input is handled properly
inputlist <- list(setA=factor(c("A", "B", "D")),
   setB=factor(c("A", "C", "E", "F")))
signed_overlaps(inputlist, return_items=TRUE)

# check to verify
signed_overlaps(inputlist, return_items=TRUE)$items

# test specific factor level order
inputlist <- list(
   setA=factor(c("A", "B", "D"), levels=c("D", "B", "A")),
   setB=factor(c("A", "C", "E", "F")))
signed_overlaps(inputlist, return_items=TRUE)

}
\seealso{
Other venndir core: 
\code{\link{render_venndir}()},
\code{\link{textvenn}()},
\code{\link{venn_meme}()},
\code{\link{venndir}()}
}
\concept{venndir core}
