% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-overlap.R
\name{signed_overlaps}
\alias{signed_overlaps}
\title{Signed overlaps}
\usage{
signed_overlaps(
  setlist,
  overlap_type = c("each", "overlap", "concordance", "concordant"),
  return_items = FALSE,
  sep = "&",
  trim_zero = TRUE,
  include_blanks = TRUE,
  ...
)
}
\arguments{
\item{setlist}{`list` of named vectors, whose names represent
set items, and whose values represent direction using 
values `c(-1, 0, 1)`.}

\item{overlap_type}{`character` value indicating the type of
overlap logic: `"each"` records each combination of
signs; `"overlap"` disregards the sign and returns any match
item overlap; `"concordance"` represents agreement in sign
for overlap, or `"mixed"` for any inconsistent overlapping
sign; `"concordant"` represents agreement in sign as
`"concordant"` which combines agree-up, and agree-down,
and any inconsistency in sign as `"mixed"`.}

\item{return_items}{`logical` indicating whether to return
the items within each overlap set.}

\item{sep}{`character` used as a delimiter between set names,
the default is `"&"`.}

\item{...}{additional arguments are passed to `list2imsigned()`.}
}
\description{
Signed overlaps
}
