% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{simple_ellipse}
\alias{simple_ellipse}
\title{Simple ellipse function}
\usage{
simple_ellipse(h, k, a, b = a, phi = 0, n = 200L)
}
\description{
Simple ellipse function
}
\seealso{
Other venndir internal: 
\code{\link{assemble_venndir_label}()},
\code{\link{expand_range}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{match_list}()},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir internal}
