% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-polyclip.R
\name{simple_ellipse}
\alias{simple_ellipse}
\title{Simple ellipse function}
\usage{
simple_ellipse(h, k, a, b = a, phi = 0, n = 200L)
}
\description{
Simple ellipse function
}
\seealso{
Other venndir polygons: 
\code{\link{get_venn_polygon_shapes}()}
}
\concept{venndir polygons}
