% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{sp_ellipses}
\alias{sp_ellipses}
\title{Make SpatialPolygons ellipses}
\usage{
sp_ellipses(
  xcenter,
  ycenter,
  setnames = NULL,
  xradius = 1,
  yradius = 2,
  rotation_degrees = c(0),
  n = 60,
  ...
)
}
\arguments{
\item{xcenter, ycenter}{\code{numeric} vector that defines the x and y
coordinate position of the center of each ellipse.}

\item{setnames}{\code{vector} that contains names for each ellipse, stored
in the \code{sp::SpatialPolygons} object as the polygon name. When
\code{setnames} is \code{NULL} then integer index positions are used.}

\item{xradius, yradius}{\code{numeric} vector that defines the radius
of each ellipse along the x-axis and y-axis, respectively.
Each \code{vector} is recycled to \code{length(xcenter)}.}

\item{n}{\code{integer} value indicating the number of subdivisions to
use in the circle.}

\item{...}{additional arguments are ignored.}

\item{rotation_degree}{\code{numeric} vector representing degrees to
rotate each ellipse after it is created, where values are
conformed to between \code{0} and \code{360}, rotating clockwise.}
}
\value{
object \code{sp::SpatialPolygons} with a number of ellipses
defined by \code{length(xcenter)}.
}
\description{
Make SpatialPolygons ellipses
}
\details{
This function creates one or more ellipses as
\code{sp::SpatialPolygons} objects.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_polylabelr}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
