% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-sp.R
\name{sp_polylabelr}
\alias{sp_polylabelr}
\title{Get SpatialPolygons polylabelr coordinate}
\usage{
sp_polylabelr(sp, apply_holes = TRUE, ...)
}
\arguments{
\item{sp}{\code{sp::SpatialPolygons} object}

\item{apply_holes}{\code{logical} indicating whether to apply any
polygon holes when present in \code{sp}.}

\item{...}{additional arguments are ignored}
}
\value{
\code{list} with items \code{x} with x coordinate, \code{y} with
y coordinate, and \code{dist} with distance to the enclosing
polygon.
}
\description{
Get SpatialPolygons polylabelr coordinate
}
\details{
This function is a simple wrapper function around
\code{polylabelr::poi()} for \code{sp::SpatialPolygons} input.
This function does two things:
\enumerate{
\item It finds the largest polygon in the \code{sp::SpatialPolygons}
input, which may contain multiple disconnected polygons.
\item It applies "holes" inside the polygon when present,
to prevent the label from being chosen inside the hole.
}
}
\examples{
setlist <- list(A=letters, B=sample(letters, 4));
vo <- venndir(setlist, proportional=TRUE, do_plot=FALSE);
sp <- vo$venn_spdf[1,];
plot(sp, col="gold")
xy <- sp_polylabelr(sp);
points(xy, pch=20, cex=4, col="navy")

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{spread_degrees}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
