% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{split_JamPolygon}
\alias{split_JamPolygon}
\title{Split JamPolygon multipart polygons}
\usage{
split_JamPolygon(jp, suffix = "_sub", ...)
}
\arguments{
\item{jp}{\code{JamPolygon}}

\item{suffix}{\code{character} passed to \code{jamba::makeNames()} when creating
new unique \code{names()} for the resulting polygons. The default \code{"_sub"}
will create sub-parts with \code{"_sub1"}, \code{"_sub2"}, \code{"_sub3"}, and so on.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{JamPolygon} with one row per distinct polygon part in the
input \code{jp} object.
}
\description{
Split JamPolygon multipart polygons into separate polygons, one
polygon per row in \code{jp@polygons}.
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{Venndir-class}},
\code{\link{[,JamPolygon-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
