% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-angles.R
\name{spread_degrees}
\alias{spread_degrees}
\title{Spread angles to minimum degrees difference}
\usage{
spread_degrees(
  degrees,
  min_degrees = 10,
  iteration = 1,
  max_iterations = 20,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{degrees}{\code{numeric} vector of angles in degrees, expected
to range from \code{0} to \code{360}. Values are fit to the range \code{c(0, 360)}
using \code{degrees \%\% 360}.}

\item{min_degrees}{\code{numeric} indicating the minimum angle in degrees
to allow between adjacent angles.}

\item{...}{additional arguments are ignored.}
}
\description{
Spread angles to minimum degrees difference
}
\details{
The function takes a vector of angles in degrees, and
adjusts any angles with adjacent angle below a minimum
angle \code{min_degrees} until the minimum angle is \code{min_degrees}.
If all input angles fit this criteria, it is returned
unchanged, otherwise it will adjust angles then
iteratively call itself until the condition is met.

If all angles are less the \code{min_degrees} degrees from
the nearest adjacent angle, then all angles are equally
spaced around 360 degrees.
}
\examples{
degrees <- c(5, 10, 15, 100, 105, 110, 200, 300, 358);
degrees
spread_degrees(degrees);

degrees2 <- sample(degrees);
degrees2
spread_degrees(degrees2);

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_largest_polygon}()},
\code{\link{get_sp_buffer}()},
\code{\link{get_venn_shapes}()},
\code{\link{intersect_polygons}()},
\code{\link{label_polygon_fill}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{nudge_sp}()},
\code{\link{nudge_venndir_label}()},
\code{\link{polygon_label_outside}()},
\code{\link{polygon_label_segment}()},
\code{\link{print_color_df}()},
\code{\link{rescale_coordinates}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{rescale_sp}()},
\code{\link{shrink_df}()},
\code{\link{sp_circles}()},
\code{\link{sp_ellipses}()},
\code{\link{sp_percent_area}()},
\code{\link{sp_polylabelr}()},
\code{\link{union_polygons}()},
\code{\link{venndir_label_style}()}
}
\concept{venndir utility}
