% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-textvenn.R
\name{textvenn}
\alias{textvenn}
\title{Text Venn diagram}
\usage{
textvenn(
  setlist,
  sets = seq_along(setlist),
  overlap_type = c("detect", "concordance", "overlap", "each", "agreement"),
  template = c("wide", "tall"),
  draw_legend = FALSE,
  set_colors = NULL,
  spacing = 5,
  padding = 1,
  inverse_title = TRUE,
  inverse_counts = FALSE,
  color_by_counts = TRUE,
  return_items = TRUE,
  unicode = TRUE,
  big.mark = ",",
  sep = "&",
  blend_preset = "ryb",
  curate_df = NULL,
  lightMode = jamba::checkLightMode(),
  debug = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{setlist}{\code{list} of item vectors; \code{list} of vectors named by item;
incidence \code{matrix} with values \code{c(0, 1)} or \code{c(FALSE, TRUE)}, or
\code{c(-1, 0, 1)}.}

\item{sets}{\code{integer} \code{vector} as index to \code{setlist}, used to
pull out a subset of the list elements. This subset is useful
because the set colors are defined for the full \code{setlist},
which allows the subset of colors to be consistent for each set.}

\item{overlap_type}{\code{character} passed to \code{signed_overlaps()}, default
'detect' will show 'concordance' when provided signed list, otherwise
it shows 'overlap' with only the counts.}

\item{template}{\code{character} string, default 'wide' indicatiing the
placement of signed counts.
\itemize{
\item 'wide' - places signed counts beside total overlap counts
\item 'tall' - places signed counts below total overlap counts
}}

\item{draw_legend}{\code{logical} default FALSE, whether to print a text
legend table below the text venn output.}

\item{set_colors}{\code{NULL} or \code{character} \code{vector} that contains
R-compatible colors. When \code{set_colors} is \code{NULL}, categorical
colors are defined using \code{colorjam::rainbowJam()}. When
\code{set_colors} is defined, the values are recycled to the total
number of sets represented by \code{setlist}.}

\item{spacing, padding}{\code{integer} values indicating the character
spacing and padding around labels printed to the console output.}

\item{inverse_title, inverse_counts}{\code{logical} indicating whether
to inverse the color, when \code{inverse_title=TRUE} then each Venn
set is printed on colored background, when \code{inverse_title=FALSE},
each set is printed with colored text with no background color.}

\item{return_items}{\code{logical} default TRUE, whether to return items
in the \code{data.frame} in column \code{"items"}.
This argument is passed to \code{signed_overlaps()}.}

\item{unicode}{\code{logical} passed to \code{curate_venn_labels()}
indicating whether the directional label can include special
Unicode characters.}

\item{big.mark}{\code{character} passed to \code{format()} for numeric labels.}

\item{blend_preset}{\code{character} string passed as \code{preset} to
\code{colorjam::blend_colors()} to define the color wheel used
during color blending operations.}

\item{curate_df}{\code{data.frame} or \code{NULL} passed to \code{curate_venn_labels()}.}

\item{lightMode}{\code{logical} default \code{jamba::checkLightMode()} checks
whether the console has a light background, and therefore needs
to have darker text. This check is incomplete, it assumes
RStudio has a light background, and everything else is dark.
To override consistently, set the option below, or add to \code{.Rprofile}:
\itemize{
\item \code{options("jam.lightMode"=TRUE)} will force lightMode=TRUE, for
light background and darker text.
\item \code{options("jam.lightMode"=FALSE)} will force lightMode=FALSE, for
dark background and lighter text.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}
}
\value{
\code{data.frame} returned using \code{invisible()}, from the
output of \code{signed_overlaps()}.
}
\description{
Text Venn diagram
}
\details{
This function is a very simple method to print a Venn diagram
using text, intended to be displayed using mono-spaced font
on an R console.
}
\examples{
# for this purpose, set lightMode=TRUE to ensure darker text
withr::with_options(list(jam.lightMode=TRUE), {

# generate test data
setlist <- make_venn_test(n_items=100, do_signed=TRUE)

# two-way Venn by default shows concordance
textvenn(setlist, sets=c(1,2))

# without signed directionality use overlap_type="overlap"
textvenn(setlist, sets=c(1,2), overlap_type="overlap")

# three-way Venn showing each signed directionality
textvenn(setlist, sets=c(1,2,3), overlap_type="each")

# larger number of items
setlist <- make_venn_test(n_items=1000000, n_sets=4, sizes=200000, do_signed=TRUE)
# text Venn with directionality
textvenn(setlist, sets=c(1,4), "agreement")

# basic text Venn with directionality
textvenn(setlist, sets=c(3,4), "each")

# simple way to show legend
venndir_legender(textvenn(setlist, sets=c(1,2,4), overlap_type="overlap"),
   combine_size=FALSE,
   legend_style="data.frame")

})
}
\seealso{
Other venndir core: 
\code{\link{render_venndir}()},
\code{\link{venn_meme}()},
\code{\link{venndir}()}
}
\concept{venndir core}
