% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{three_point_angle}
\alias{three_point_angle}
\title{Calculate angle between three consecutive points}
\usage{
three_point_angle(x)
}
\arguments{
\item{x}{\code{matrix} with first two columns assumed to contain \code{x}, \code{y}
coordinates.}
}
\value{
\code{numeric} angle defined by the three points, where
the second point is considered the vertex.
}
\description{
Calculate angle between three consecutive points
}
\examples{
x <- cbind(x=c(1, 1, 3), y=c(2, 1, 1))
three_point_angle(x)

}
\seealso{
Other venndir geometry: 
\code{\link{degrees_to_adj}()},
\code{\link{diff_degrees}()},
\code{\link{display_angles}()},
\code{\link{mean_degree_arc}()},
\code{\link{mean_degrees}()},
\code{\link{rescale_coordinates}()},
\code{\link{spread_degrees}()}
}
\concept{venndir geometry}
