% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JamPolygon.R
\name{three_point_angle}
\alias{three_point_angle}
\title{Calculate angle between three consecutive points}
\usage{
three_point_angle(x)
}
\arguments{
\item{x}{\code{matrix} with first two columns assumed to contain \code{x}, \code{y}
coordinates.}
}
\description{
Calculate angle between three consecutive points
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{expand_range}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{modify_venndir_overlap}()},
\code{\link{nudge_venndir_label}()},
\code{\link{plot,Venndir,ANY-method}},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{venndir_legender}()},
\code{\link{venndir_to_df}()}
}
\concept{venndir utility}
