% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndir-class.R
\docType{class}
\name{Venndir-class}
\alias{Venndir-class}
\title{Venndir class}
\description{
Venndir class
}
\details{
This object is intended to contain all the required data to produce
a venndir figure. The components in this object can be edited.

Slots:
\itemize{
\item \strong{jps}: \code{JamPolygon} containing the set and overlap polygons.
\itemize{
\item Columns \code{"x"},\code{"y"} contain polygon coordinates, see
\code{JamPolygon-class} for details on multi-part polygons, and polygons
with holes.
\item The polygon type is defined with column \code{type} and values
\code{"set"} or \code{"overlap"}.
\item The "set" polygons are intended to be the full circle for each
set, but may be \code{NA} or \code{NULL} to indicate the set is empty.
\item The "overlap" polygons are expected to be contained inside
"set" polygons, and each overlap polygon may be multi-polygons
with or without holes.
For example, the "set" polygon for \code{set_A} which is a circle fully
contained inside another set \code{set_B}.
The "overlap" polygon for \code{set_A} would be empty, since there is no
region unique to \code{set_A}.
However the "overlap" polygon \code{set_A&set_B}
would contain the full circle defined by "set" polygon \code{set_A}.
Lastly the "overlap" polygon for \code{set_B} would be a circle with a hole
located at "overlap" \code{set_A&set_B}.
or ellipse. The overlap polygons are polygons, possibly multi-part
with or without holes, or may be \code{NA} or \code{NULL} to indicate that
this overlap does not have a corresponding polygon.
\item The \code{name} column contains set or overlap
names, where multiple sets are concatenated with \code{"&"} by default.
or ellipse corresponding to each set in \code{setlist}), and the overlap
polygons. Where an overlap does not exist, the polygon coordinates
will be empty, or will have entirely \code{NA} values for "x" and "y".
}
\item \strong{label_df}: \code{data.frame} which contains information about
Venn labels placement and content.
It should contain one row for each Venn overlap.
When data contains signed overlaps, each overlap is sub-divided
by the combination of overlap signs, with one row per overlap and
overlap signs.
Overlap signs are summarized based upon argument \code{overlap_type},
see \code{signed_overlaps()}.
\itemize{
\item \code{"overlap_sign"} represents the combination of overlap name,
and the overlap sign across all sets.
\item \code{"text"} indicates the count label to be displayed.
\item \code{"venn_counts"} indicates the numeric number of Venn overlaps
in each row.
\item \code{"item"} should be a \code{list} that contains \code{character} vectors
of items in each overlap set. Column \code{"item"} is optional.
\item \code{"nsets"} indicates the number of sets involved in the overlap.
\item \code{"x"}, \code{"y"} represent the label coordinate to use inside
each polygon overlap. When the label is displayed outside the polygon,
the values \code{"x_offset"} and \code{"y_offset"} are added to \code{"x"} and \code{"y"},
respectively. In this way, a line segment can be drawn from the outer
label back to the corresponding polygon.
\item \code{"color"} indicates the color for each label. Note the
actual color may be modified based upon the background fill color
to maximize the contrast and visibility of the color. For example
blue on a dark background is displayed as light blue.
\item \code{"fontsize"} indicates the base font size, prior to any
optional adjustment with \code{font_cex}.
\item \code{"fill"} indicates an optional color fill behind the count
label. When \code{NULL} (default) there is no color fill. This value
is usually populated by argument \code{label_style} which is passed to
\code{venndir_label_preset()}.
\item \code{"border"} indicates an optional border around count labels,
and is handled similar to \code{"fill"}.
\item \code{"padding"}, \code{"padding_unit"}, \code{"r"}, and \code{"r_unit"} are
used with \code{gridtext::richtext_grob()} to define padding around count
labels, and optional rounded corners when border is displayed.
\item Other columns are intended for internal use by \code{render_venndir()}
and are not yet fully editable.
}
\item \strong{setlist}: \code{list} with the \code{setlist} used to create the Venn overlaps.
Previously this data could be inferred from \code{label_df} which was
tedious, and required column \code{"item"} which is optional.
That said, \code{setlist} can be an empty \code{list()}.
\item \strong{metadata}: \code{list} with optional metadata, intended for future
expansion, such as plot title.
}
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{[,JamPolygon-method}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{nudge_JamPolygon}()},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
