% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndir-class.R
\docType{class}
\name{Venndir-class}
\alias{Venndir-class}
\title{Venndir class}
\description{
Venndir class
}
\details{
This object is intended to contain all the required data to produce
a venndir figure. The components in this object can be edited.

Slots:
\itemize{
\item \strong{jps}: \code{JamPolygon} containing the set polygons (usually the full circle
or ellipse corresponding to each set in \code{setlist}), and the overlap
polygons. Where an overlap does not exist, the polygon coordinates
will be empty, or will have entirely \code{NA} values for "x" and "y".
\item \strong{label_df}: \code{data.frame} containing detailed information about
where to place Venn labels in the figure, what font to use, color,
and items (when items are necessary for the figure). The "x" and "y"
coordinates define the location of each label inside the polygon,
the "x_offset" and "y_offset" are added to those coordinates to
define the label position outside the polygon.
\item \strong{setlist}: \code{list} with the \code{setlist} used to create the Venn overlaps.
Previously this data could be inferred from \code{label_df} which was tedious.
}
}
\seealso{
Other JamPolygon: 
\code{\link{JamPolygon-class}},
\code{\link{add_orientation_JamPolygon}()},
\code{\link{area_JamPolygon}()},
\code{\link{bbox_JamPolygon}()},
\code{\link{buffer_JamPolygon}()},
\code{\link{check_JamPolygon}()},
\code{\link{check_Venndir}()},
\code{\link{eulerr_to_JamPolygon}()},
\code{\link{farthest_point_JamPolygon}()},
\code{\link{find_venn_overlaps_JamPolygon}()},
\code{\link{has_point_in_JamPolygon}()},
\code{\link{intersect_JamPolygon}()},
\code{\link{label_fill_JamPolygon}()},
\code{\link{label_outside_JamPolygon}()},
\code{\link{label_segment_JamPolygon}()},
\code{\link{labelr_JamPolygon}()},
\code{\link{minus_JamPolygon}()},
\code{\link{nearest_point_JamPolygon}()},
\code{\link{plot,JamPolygon,missing-method}},
\code{\link{plot.JamPolygon}()},
\code{\link{point_in_JamPolygon}()},
\code{\link{polyclip_to_JamPolygon}()},
\code{\link{sample_JamPolygon}()},
\code{\link{split_JamPolygon}()},
\code{\link{union_JamPolygon}()},
\code{\link{update_JamPolygon}()}
}
\concept{JamPolygon}
