% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndir-class.R
\docType{methods}
\name{plot,Venndir,ANY-method}
\alias{plot,Venndir,ANY-method}
\alias{length,Venndir-method}
\alias{setlist,Venndir-method}
\alias{names,Venndir-method}
\alias{show,Venndir-method}
\alias{metadata,Venndir-method}
\alias{metadata<-,Venndir-method}
\alias{overlaplist,Venndir-method}
\alias{overlapdf,Venndir-method}
\alias{signed_counts,Venndir-method}
\alias{im,Venndir-method}
\title{Plot Venndir object}
\usage{
\S4method{plot}{Venndir,ANY}(x, y, ...)

\S4method{length}{Venndir}(x)

\S4method{setlist}{Venndir}(x)

\S4method{names}{Venndir}(x)

\S4method{show}{Venndir}(object)

\S4method{metadata}{Venndir}(x)

\S4method{metadata}{Venndir}(x) <- value

\S4method{overlaplist}{Venndir}(x)

\S4method{overlapdf}{Venndir}(x)

\S4method{signed_counts}{Venndir}(x)

\S4method{im}{Venndir}(x)
}
\value{
\code{Venndir} object, invisibly.
}
\description{
Plot Venndir object
}
\seealso{
Other venndir core: 
\code{\link{make_venn_test}()},
\code{\link{render_venndir}()},
\code{\link{signed_overlaps}()},
\code{\link{textvenn}()},
\code{\link{venn_meme}()},
\code{\link{venndir}()},
\code{\link{venndir_legender}()}
}
\concept{venndir core}
