% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-render.R
\name{venndir_label_style}
\alias{venndir_label_style}
\title{venndir label style}
\usage{
venndir_label_style(
  venndir_output,
  label_preset = c("none", "main inside", "main outside", "outside", "all outside",
    "concept", "meme", "items", "main items", "main count items", "custom"),
  label_style = c("basic", "box", "fill", "fill_box", "shaded", "shaded_box", "lite",
    "lite_box"),
  lite = "#FFEEAABB",
  bg = "white",
  set = c("outside", "inside", "none"),
  overlap = c("none", "inside", "outside"),
  count = c("inside", "outside", "ifneeded", "detect", "none"),
  signed = c("inside", "outside", "ifneeded", "detect", "none"),
  items = c("none", "inside"),
  max_items = 3000,
  inside_percent_threshold = 5,
  label_types = c("main", "signed"),
  show_zero = TRUE,
  sep = "&",
  useGrey = 15,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{label_style}{\code{character} string indicating the style of label
to display. The values \verb{"basic","none","shaded","lite","fill"}
style the label background fill, while presence of \code{"box"} in
the string will draw a border around the label:
\code{"basic"} or \code{"none"} uses no background fill,
\code{"lite"} uses lite background fill,
\code{"fill"} uses opaque fill with the overlap set color,
\code{"shaded"} uses slightly transparent fill with overlap set color,
\code{"box"} displays border around the label.}
}
\description{
venndir label style
}
\details{
This function applies a label style to \code{venndir()} output,
overwriting the existing label style as necessary.

This function adjusts the label text color for contrast with
\code{make_color_contrast()}, which is useful when positioning
the label on top of dark or bright colors. If the label
is positioned outside the \code{sp} polygons, the text is assumed
to be on a white background, using argument \code{bg}.
}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{eulerr2polys}()},
\code{\link{expand_range}()},
\code{\link{find_vennpoly_overlaps}()},
\code{\link{get_venn_shapes}()},
\code{\link{gridtext_make_outer_box}()},
\code{\link{gridtext_richtext_grob}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{nudge_venndir_label}()},
\code{\link{print_color_df}()},
\code{\link{rescale_ps}()},
\code{\link{rescale_p}()},
\code{\link{shrink_df}()},
\code{\link{sp_percent_area}()},
\code{\link{venndir_legender}()}
}
\concept{venndir utility}
