% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-to-df.R
\name{venndir_to_df}
\alias{venndir_to_df}
\title{Convert venndir output to data.frame}
\usage{
venndir_to_df(
  venndir_out,
  df_format = c("hits", "items", "wide"),
  return_type = c("data.frame", "kable"),
  trim_blanks = TRUE,
  wrap_width = 80,
  colorize_headers = TRUE,
  set_colors = NULL,
  item_type = "gene",
  add_counts = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{venndir_out}{\code{list} output from \code{venndir()}}

\item{df_format}{\code{character} string, default "hits" with output format:
\itemize{
\item \code{"hits"} - returns a hit matrix, with the first column containing
item labels, subsequent columns indicate "hits" with 0 or +1 or -1.
\item \code{"items"} - returns one column for each overlap, containing
item labels in each corresponding column. This output does not
include the sign.
\item \code{"wide"} - returns one row for each overlap_set/sign combination,
and a column \code{"items"} which is a \code{list} of \code{character} vectors
with the item labels. Rows may be split for "word wrapping"
when there are too many items. This output is intended for
\code{kable()} formatting with grouped rows, for example with argument
\code{return_type="kable"}.
}}

\item{return_type}{\code{character} string indicating how to format output:
\itemize{
\item \code{"kable"}: returns an object \code{"knitr_kable"} suitable for printing
as HTML or LaTeX. This output is recommended for RMarkdown documents.
\item \code{"data.frame"}: returns a \code{data.frame} object without colorization.
}}

\item{trim_blanks}{\code{logical} whether to remove rows with 0 items.}

\item{wrap_width}{\code{numeric} maximum width for item text before it
is word-wrapped. To disable word-wrapping use \code{wrap_width=Inf}.}

\item{colorize_headers}{\code{logical} indicating whether to color the
background of each header by the \code{set_colors}.}

\item{...}{additional arguments are ignored.}
}
\value{
object of class \code{"data.frame"} or \code{"knitr_kable"} based upon
argument \code{return_type}.
}
\description{
Convert venndir output to data.frame or kable format
for visual review.
}
\details{
\subsection{Todo:}{
\itemize{
\item Accept input from \code{signed_overlaps()} for purely table summary.
}
}
}
\examples{
setlist <- venndir::make_venn_test(100, 3, do_signed=TRUE);
venndir_out <- venndir::venndir(setlist, overlap_type="each")
df <- venndir_to_df(venndir_out)
head(df, 10)

kdf <- venndir_to_df(venndir_out, return_type="kable")
kdf

df2 <- venndir_to_df(venndir_out, df_format="items")
head(df2, 10)

kdf2 <- venndir_to_df(venndir_out, df_format="items", return_type="kable")
kdf2

df3 <- venndir_to_df(venndir_out, df_format="wide", return_type="data.frame")
df3

kdf3 <- venndir_to_df(venndir_out, df_format="wide", return_type="kable")
kdf3

}
\seealso{
Other venndir utility: 
\code{\link{curate_venn_labels}()},
\code{\link{expand_range}()},
\code{\link{get_venndir_curate_df}()},
\code{\link{make_color_contrast}()},
\code{\link{make_venn_combn_df}()},
\code{\link{make_venn_test}()},
\code{\link{match_list}()},
\code{\link{modify_venndir_overlap}()},
\code{\link{nudge_venndir_label}()},
\code{\link{plot,Venndir,ANY-method}},
\code{\link{print_color_df}()},
\code{\link{shrink_df}()},
\code{\link{three_point_angle}()},
\code{\link{venndir_legender}()}
}
\concept{venndir utility}
