% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndir-poly-conversions.R
\name{xy_list_to_polygon_list}
\alias{xy_list_to_polygon_list}
\title{Convert coordinate list of x and y into polygon list of x,y coordinates}
\usage{
xy_list_to_polygon_list(x, ...)
}
\description{
Convert coordinate list of x and y into polygon list of x,y coordinates
}
\examples{
counts <- c(A=1, B=2, `A&B`=3)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)
xy_list <- polygon_list_to_xy_list(polygon_list)
polygon_list2 <- xy_list_to_polygon_list(xy_list)
identical(polygon_list, polygon_list2)

}
\seealso{
Other venndir polygons: 
\code{\link{bbox_polygon_list}()},
\code{\link{eulerr_to_polygon_list}()},
\code{\link{get_largest_polygon_list}()},
\code{\link{get_venn_polygon_shapes}()},
\code{\link{intersect_polygon_list}()},
\code{\link{labelr_polygon_list}()},
\code{\link{minus_polygon_list}()},
\code{\link{nudge_polygon_coords}()},
\code{\link{nudge_polygon_list}()},
\code{\link{plot_polygon_list}()},
\code{\link{polygon_areas}()},
\code{\link{polygon_circles}()},
\code{\link{polygon_ellipses}()},
\code{\link{polygon_list_labelr}()},
\code{\link{polygon_list_to_xy_list}()},
\code{\link{rescale_polygon_list}()},
\code{\link{simple_ellipse}()},
\code{\link{union_polygon_list}()}
}
\concept{venndir polygons}
